split commands.go and remove lots of dead code
