+++ pkg/cli/commands.go
@@ -1,28 +1,14 @@
 package cli
 
 import (
-	"bufio"
 	_ "embed"
-	"encoding/json"
 	"fmt"
-	"math/rand"
 	"os"
 	"os/exec"
-	"os/signal"
 	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
-	"syscall"
-	"time"
 
-	"github.com/cli/go-gh/v2"
-	"github.com/fsnotify/fsnotify"
-	"github.com/githubnext/gh-aw/pkg/console"
 	"github.com/githubnext/gh-aw/pkg/constants"
-	"github.com/githubnext/gh-aw/pkg/parser"
-	"github.com/githubnext/gh-aw/pkg/workflow"
-	"github.com/goccy/go-yaml"
 )
 
 // Package-level version information
@@ -46,2582 +32,9 @@ func GetVersion() string {
 	return version
 }
 
-// GitHubWorkflow represents a GitHub Actions workflow from the API
-// GitHubWorkflowsResponse represents the GitHub API response for workflows
-// Note: The API returns an array directly, not wrapped in a workflows field
-
-// ListEnginesAndOtherInformation lists available workflow components
-func ListEnginesAndOtherInformation(verbose bool) error {
-	if verbose {
-		fmt.Fprintln(os.Stderr, console.FormatProgressMessage("Searching for available workflow components..."))
-	}
-
-	// List available agentic engines
-	if err := listAgenticEngines(verbose); err != nil {
-		fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to list agentic engines: %v", err)))
-	}
-
-	// Provide information about workflow repositories
-	fmt.Println("\nTo add workflows to your project:")
-	fmt.Println("=================================")
-	fmt.Println("Use the 'add' command with repository/workflow specifications:")
-	fmt.Println("  " + constants.CLIExtensionPrefix + " add owner/repo/workflow-name")
-	fmt.Println("  " + constants.CLIExtensionPrefix + " add owner/repo/workflow-name@version")
-	fmt.Println("\nExample:")
-	fmt.Println("  " + constants.CLIExtensionPrefix + " add githubnext/agentics/ci-doctor")
-	fmt.Println("  " + constants.CLIExtensionPrefix + " add githubnext/agentics/daily-plan@main")
-	return nil
-}
-
-// listAgenticEngines lists all available agentic engines with their characteristics
-func listAgenticEngines(verbose bool) error {
-	// Create an engine registry directly to access the engines
-	registry := workflow.GetGlobalEngineRegistry()
-
-	// Get all supported engines from the registry
-	engines := registry.GetSupportedEngines()
-
-	if len(engines) == 0 {
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage("No agentic engines available."))
-		return nil
-	}
-
-	// Build table configuration
-	var headers []string
-	if verbose {
-		headers = []string{"ID", "Display Name", "Status", "MCP", "HTTP Transport", "Description"}
-	} else {
-		headers = []string{"ID", "Display Name", "Status", "MCP", "HTTP Transport"}
-	}
-
-	var rows [][]string
-
-	for _, engineID := range engines {
-		engine, err := registry.GetEngine(engineID)
-		if err != nil {
-			if verbose {
-				fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to get engine '%s': %v", engineID, err)))
-			}
-			continue
-		}
-
-		// Determine status
-		status := "Stable"
-		if engine.IsExperimental() {
-			status = "Experimental"
-		}
-
-		// MCP support
-		mcpSupport := "No"
-		if engine.SupportsToolsAllowlist() {
-			mcpSupport = "Yes"
-		}
-
-		// HTTP transport support
-		httpTransport := "No"
-		if engine.SupportsHTTPTransport() {
-			httpTransport = "Yes"
-		}
-
-		// Build row data
-		var row []string
-		if verbose {
-			row = []string{
-				engine.GetID(),
-				engine.GetDisplayName(),
-				status,
-				mcpSupport,
-				httpTransport,
-				engine.GetDescription(),
-			}
-		} else {
-			row = []string{
-				engine.GetID(),
-				engine.GetDisplayName(),
-				status,
-				mcpSupport,
-				httpTransport,
-			}
-		}
-		rows = append(rows, row)
-	}
-
-	// Render the table
-	tableConfig := console.TableConfig{
-		Title:   "Available Agentic Engines",
-		Headers: headers,
-		Rows:    rows,
-	}
-	fmt.Fprint(os.Stderr, console.RenderTable(tableConfig))
-
-	fmt.Fprintln(os.Stderr, "")
-	return nil
-}
-
-// AddWorkflows adds one or more workflows from components to .github/workflows
-// with optional repository installation and PR creation
-func AddWorkflows(workflows []string, number int, verbose bool, engineOverride string, repoSpec string, name string, force bool, createPR bool) error {
-	if len(workflows) == 0 {
-		return fmt.Errorf("at least one workflow name is required")
-	}
-
-	for i, workflow := range workflows {
-		if workflow == "" {
-			return fmt.Errorf("workflow name cannot be empty (workflow %d)", i+1)
-		}
-	}
-
-	// The -r flag is no longer supported
-	if repoSpec != "" {
-		return fmt.Errorf("-r flag is deprecated; use the new format: owner/repo/workflow-name[@version]")
-	}
-
-	// If creating a PR, check prerequisites
-	if createPR {
-		// Check if GitHub CLI is available
-		if !isGHCLIAvailable() {
-			return fmt.Errorf("GitHub CLI (gh) is required for PR creation but not available")
-		}
-
-		// Check if we're in a git repository
-		if !isGitRepo() {
-			return fmt.Errorf("not in a git repository - PR creation requires a git repository")
-		}
-
-		// Check no other changes are present
-		if err := checkCleanWorkingDirectory(verbose); err != nil {
-			return fmt.Errorf("working directory is not clean: %w", err)
-		}
-	}
-
-	// Parse workflow specifications and group by repository
-	repoToWorkflows := make(map[string][]string) // repo -> list of workflow names
-	repoVersions := make(map[string]string)      // repo -> version
-	processedWorkflows := make([]string, 0, len(workflows))
-
-	for _, workflow := range workflows {
-		spec, err := parseWorkflowSpec(workflow)
-		if err != nil {
-			return fmt.Errorf("invalid workflow specification '%s': %w", workflow, err)
-		}
-
-		// Handle repository installation and workflow name extraction
-		if existing, exists := repoVersions[spec.Repo]; exists && existing != spec.Version {
-			return fmt.Errorf("conflicting versions for repository %s: %s vs %s", spec.Repo, existing, spec.Version)
-		}
-		repoVersions[spec.Repo] = spec.Version
-		repoToWorkflows[spec.Repo] = append(repoToWorkflows[spec.Repo], spec.WorkflowPath)
-
-		// Create qualified name for processing
-		processedWorkflows = append(processedWorkflows, fmt.Sprintf("%s/%s", spec.Repo, spec.WorkflowPath))
-	}
-
-	// Install required repositories
-	for repo, version := range repoVersions {
-		repoWithVersion := repo
-		if version != "" {
-			repoWithVersion = fmt.Sprintf("%s@%s", repo, version)
-		}
-
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Installing repository %s before adding workflows...", repoWithVersion)))
-		}
-
-		// Install as global package (not local) to match the behavior expected
-		if err := InstallPackage(repoWithVersion, verbose); err != nil {
-			return fmt.Errorf("failed to install repository %s: %w", repoWithVersion, err)
-		}
-	}
-
-	// Handle PR creation workflow
-	if createPR {
-		return addWorkflowsWithPR(processedWorkflows, number, verbose, engineOverride, name, force)
-	}
-
-	// Handle normal workflow addition
-	return addWorkflowsNormal(processedWorkflows, number, verbose, engineOverride, name, force)
-}
-
-// addWorkflowsNormal handles normal workflow addition without PR creation
-func addWorkflowsNormal(workflows []string, number int, verbose bool, engineOverride string, name string, force bool) error {
-	// Create file tracker for all operations
-	tracker, err := NewFileTracker()
-	if err != nil {
-		// If we can't create a tracker (e.g., not in git repo), fall back to non-tracking behavior
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Could not create file tracker: %v", err)))
-		}
-		tracker = nil
-	}
-
-	if len(workflows) > 1 {
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Adding %d workflow(s)...", len(workflows))))
-	}
-
-	// Add each workflow
-	for i, workflow := range workflows {
-		if len(workflows) > 1 {
-			fmt.Fprintln(os.Stderr, console.FormatProgressMessage(fmt.Sprintf("Adding workflow %d/%d: %s", i+1, len(workflows), workflow)))
-		}
-
-		// For multiple workflows, only use the name flag for the first one
-		currentName := ""
-		if i == 0 && name != "" {
-			currentName = name
-		}
-
-		if err := AddWorkflowWithTracking(workflow, number, verbose, engineOverride, currentName, force, tracker); err != nil {
-			return fmt.Errorf("failed to add workflow '%s': %w", workflow, err)
-		}
-	}
-
-	if len(workflows) > 1 {
-		fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Successfully added all %d workflows", len(workflows))))
-	}
-
-	return nil
-}
-
-// addWorkflowsWithPR handles workflow addition with PR creation
-func addWorkflowsWithPR(workflows []string, number int, verbose bool, engineOverride string, name string, force bool) error {
-	// Get current branch for restoration later
-	currentBranch, err := getCurrentBranch()
-	if err != nil {
-		return fmt.Errorf("failed to get current branch: %w", err)
-	}
-
-	// Create temporary branch with random 4-digit number
-	randomNum := rand.Intn(9000) + 1000 // Generate number between 1000-9999
-	var branchName string
-	if len(workflows) == 1 {
-		branchName = fmt.Sprintf("add-workflow-%s-%04d", strings.ReplaceAll(workflows[0], "/", "-"), randomNum)
-	} else {
-		workflowNames := strings.Join(workflows, "-")
-		if len(workflowNames) > 50 { // Truncate long branch names
-			workflowNames = workflowNames[:47] + "..."
-		}
-		branchName = fmt.Sprintf("add-workflows-%s-%04d", strings.ReplaceAll(workflowNames, "/", "-"), randomNum)
-	}
-
-	if err := createAndSwitchBranch(branchName, verbose); err != nil {
-		return fmt.Errorf("failed to create branch %s: %w", branchName, err)
-	}
-
-	// Create file tracker for rollback capability
-	tracker, err := NewFileTracker()
-	if err != nil {
-		return fmt.Errorf("failed to create file tracker: %w", err)
-	}
-
-	// Ensure we switch back to original branch on exit
-	defer func() {
-		if switchErr := switchBranch(currentBranch, verbose); switchErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to switch back to branch %s: %v", currentBranch, switchErr)))
-		}
-	}()
-
-	// Add workflows using the normal function logic
-	if err := addWorkflowsNormal(workflows, number, verbose, engineOverride, name, force); err != nil {
-		// Rollback on error
-		if rollbackErr := tracker.RollbackAllFiles(verbose); rollbackErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to rollback files: %v", rollbackErr)))
-		}
-		return fmt.Errorf("failed to add workflows: %w", err)
-	}
-
-	// Stage all files before creating PR
-	if err := tracker.StageAllFiles(verbose); err != nil {
-		if rollbackErr := tracker.RollbackAllFiles(verbose); rollbackErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to rollback files: %v", rollbackErr)))
-		}
-		return fmt.Errorf("failed to stage workflow files: %w", err)
-	}
-
-	// Update .gitattributes and stage it if modified
-	if err := stageGitAttributesIfChanged(); err != nil && verbose {
-		fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to stage .gitattributes: %v", err)))
-	}
-
-	// Commit changes
-	var commitMessage, prTitle, prBody string
-	if len(workflows) == 1 {
-		commitMessage = fmt.Sprintf("Add workflow: %s", workflows[0])
-		prTitle = fmt.Sprintf("Add workflow: %s", workflows[0])
-		prBody = fmt.Sprintf("Automatically created PR to add workflow: %s", workflows[0])
-	} else {
-		commitMessage = fmt.Sprintf("Add workflows: %s", strings.Join(workflows, ", "))
-		prTitle = fmt.Sprintf("Add workflows: %s", strings.Join(workflows, ", "))
-		prBody = fmt.Sprintf("Automatically created PR to add workflows: %s", strings.Join(workflows, ", "))
-	}
-
-	if err := commitChanges(commitMessage, verbose); err != nil {
-		if rollbackErr := tracker.RollbackAllFiles(verbose); rollbackErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to rollback files: %v", rollbackErr)))
-		}
-		return fmt.Errorf("failed to commit files: %w", err)
-	}
-
-	// Push branch
-	if err := pushBranch(branchName, verbose); err != nil {
-		if rollbackErr := tracker.RollbackAllFiles(verbose); rollbackErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to rollback files: %v", rollbackErr)))
-		}
-		return fmt.Errorf("failed to push branch %s: %w", branchName, err)
-	}
-
-	// Create PR
-	if err := createPR(branchName, prTitle, prBody, verbose); err != nil {
-		if rollbackErr := tracker.RollbackAllFiles(verbose); rollbackErr != nil && verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to rollback files: %v", rollbackErr)))
-		}
-		return fmt.Errorf("failed to create PR: %w", err)
-	}
-
-	// Success - no rollback needed
-
-	// Switch back to original branch
-	if err := switchBranch(currentBranch, verbose); err != nil {
-		return fmt.Errorf("failed to switch back to branch %s: %w", currentBranch, err)
-	}
-
-	if len(workflows) == 1 {
-		fmt.Printf("Successfully created PR for workflow: %s\n", workflows[0])
-	} else {
-		fmt.Printf("Successfully created PR for workflows: %s\n", strings.Join(workflows, ", "))
-	}
-	return nil
-}
-
-// Legacy function wrappers for backwards compatibility
-// These can be removed once all tests are updated
-
-// AddWorkflowWithRepo adds a workflow from components to .github/workflows
-// with optional repository installation
-// Deprecated: Use AddWorkflows instead
-func AddWorkflowWithRepo(workflow string, number int, verbose bool, engineOverride string, repoSpec string, name string, force bool) error {
-	// Handle empty workflow name like the original function (show help, don't error)
-	if workflow == "" {
-		fmt.Fprintln(os.Stderr, console.FormatErrorMessage("No components path specified. Usage: "+constants.CLIExtensionPrefix+" add <name>"))
-		// Show available workflows using the same logic as ListEnginesAndOtherInformation
-		return ListEnginesAndOtherInformation(false)
-	}
-	return AddWorkflows([]string{workflow}, number, verbose, engineOverride, repoSpec, name, force, false)
-}
-
-// AddWorkflowWithRepoAndPR adds a workflow from components to .github/workflows
-// with optional repository installation and creates a PR
-// Deprecated: Use AddWorkflows instead
-func AddWorkflowWithRepoAndPR(workflow string, number int, verbose bool, engineOverride string, repoSpec string, name string, force bool) error {
-	// Handle empty workflow name like the original function (show help, don't error)
-	if workflow == "" {
-		fmt.Fprintln(os.Stderr, console.FormatErrorMessage("No components path specified. Usage: "+constants.CLIExtensionPrefix+" add <name>"))
-		// Show available workflows using the same logic as ListEnginesAndOtherInformation
-		return ListEnginesAndOtherInformation(false)
-	}
-	return AddWorkflows([]string{workflow}, number, verbose, engineOverride, repoSpec, name, force, true)
-}
-
-// AddMultipleWorkflowsWithRepo adds multiple workflows from components to .github/workflows
-// with optional repository installation
-// Deprecated: Use AddWorkflows instead
-func AddMultipleWorkflowsWithRepo(workflows []string, number int, verbose bool, engineOverride string, repoSpec string, name string, force bool) error {
-	return AddWorkflows(workflows, number, verbose, engineOverride, repoSpec, name, force, false)
-}
-
-// AddMultipleWorkflowsWithRepoAndPR adds multiple workflows from components to .github/workflows
-// with optional repository installation and creates a PR
-// Deprecated: Use AddWorkflows instead
-func AddMultipleWorkflowsWithRepoAndPR(workflows []string, number int, verbose bool, engineOverride string, repoSpec string, name string, force bool) error {
-	return AddWorkflows(workflows, number, verbose, engineOverride, repoSpec, name, force, true)
-}
-
-// AddWorkflowWithTracking adds a workflow from components to .github/workflows with file tracking
-func AddWorkflowWithTracking(workflow string, number int, verbose bool, engineOverride string, name string, force bool, tracker *FileTracker) error {
-	if workflow == "" {
-		fmt.Fprintln(os.Stderr, console.FormatErrorMessage("No components path specified. Usage: "+constants.CLIExtensionPrefix+" add <name>"))
-		// Show available workflows using the same logic as ListEnginesAndOtherInformation
-		return ListEnginesAndOtherInformation(false)
-	}
-
-	if verbose {
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Adding workflow: %s", workflow)))
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Number of copies: %d", number)))
-		if force {
-			fmt.Fprintln(os.Stderr, console.FormatInfoMessage("Force flag enabled: will overwrite existing files"))
-		}
-	}
-
-	// Validate number of copies
-	if number < 1 {
-		return fmt.Errorf("number of copies must be a positive integer")
-	}
-
-	if verbose {
-		fmt.Fprintln(os.Stderr, "Locating workflow components...")
-	}
-
-	// Add .md extension if not present
-	workflowPath := workflow
-	if !strings.HasSuffix(workflowPath, ".md") {
-		workflowPath += ".md"
-	}
-
-	if verbose {
-		fmt.Printf("Looking for workflow file: %s\n", workflowPath)
-	}
-
-	// Try to read the workflow content from multiple sources
-	sourceContent, sourceInfo, err := findWorkflowInPackages(workflowPath, verbose)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, console.FormatErrorMessage(fmt.Sprintf("Workflow '%s' not found.", workflow)))
-
-		// Show available workflows using the same logic as ListEnginesAndOtherInformation
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage("Run '"+constants.CLIExtensionPrefix+" list' to see available workflows."))
-
-		return fmt.Errorf("workflow not found: %s", workflow)
-	}
-
-	if verbose {
-		fmt.Printf("Successfully read workflow content (%d bytes)\n", len(sourceContent))
-	}
-
-	// Find git root to ensure consistent placement
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return fmt.Errorf("add workflow requires being in a git repository: %w", err)
-	}
-
-	// Ensure .github/workflows directory exists relative to git root
-	githubWorkflowsDir := filepath.Join(gitRoot, ".github/workflows")
-	if err := os.MkdirAll(githubWorkflowsDir, 0755); err != nil {
-		return fmt.Errorf("failed to create .github/workflows directory: %w", err)
-	}
-
-	// Determine the filename to use
-	var filename string
-	if name != "" {
-		// Use the explicitly provided name
-		filename = name
-	} else {
-		// Extract filename from workflow path and remove .md extension for processing
-		filename = filepath.Base(workflow)
-		filename = strings.TrimSuffix(filename, ".md")
-	}
-
-	// Check if a workflow with this name already exists
-	existingFile := filepath.Join(githubWorkflowsDir, filename+".md")
-	if _, err := os.Stat(existingFile); err == nil && !force {
-		return fmt.Errorf("workflow '%s' already exists in .github/workflows/. Use a different name with -n flag, remove the existing workflow first, or use --force to overwrite", filename)
-	}
-
-	// Collect all @include dependencies from the workflow file
-	includeDeps, err := collectPackageIncludeDependencies(string(sourceContent), sourceInfo.PackagePath, verbose)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to collect include dependencies: %v", err)))
-	}
-
-	// Copy all @include dependencies to .github/workflows maintaining relative paths
-	if err := copyIncludeDependenciesFromPackageWithForce(includeDeps, githubWorkflowsDir, verbose, force, tracker); err != nil {
-		fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to copy include dependencies: %v", err)))
-	}
-
-	// Process each copy
-	for i := 1; i <= number; i++ {
-		// Construct the destination file path with numbering in .github/workflows
-		var destFile string
-		if number == 1 {
-			destFile = filepath.Join(githubWorkflowsDir, filename+".md")
-		} else {
-			destFile = filepath.Join(githubWorkflowsDir, fmt.Sprintf("%s-%d.md", filename, i))
-		}
-
-		// Check if destination file already exists
-		fileExists := false
-		if _, err := os.Stat(destFile); err == nil {
-			fileExists = true
-			if !force {
-				fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Destination file '%s' already exists, skipping.", destFile)))
-				continue
-			}
-			fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Overwriting existing file: %s", destFile)))
-		}
-
-		// Process content for numbered workflows
-		content := string(sourceContent)
-		if number > 1 {
-			// Update H1 title to include number
-			content = updateWorkflowTitle(content, i)
-		}
-
-		// Track the file based on whether it existed before (if tracker is available)
-		if tracker != nil {
-			if fileExists {
-				tracker.TrackModified(destFile)
-			} else {
-				tracker.TrackCreated(destFile)
-			}
-		}
-
-		// Write the file
-		if err := os.WriteFile(destFile, []byte(content), 0644); err != nil {
-			return fmt.Errorf("failed to write destination file '%s': %w", destFile, err)
-		}
-
-		fmt.Printf("Added workflow: %s\n", destFile)
-
-		// Try to compile the workflow and track generated files
-		if tracker != nil {
-			if err := compileWorkflowWithTracking(destFile, verbose, engineOverride, tracker); err != nil {
-				fmt.Fprintln(os.Stderr, err)
-			}
-		} else {
-			// Fall back to basic compilation without tracking
-			if err := compileWorkflow(destFile, verbose, engineOverride); err != nil {
-				fmt.Fprintln(os.Stderr, err)
-			}
-		}
-	}
-
-	// Stage tracked files to git if in a git repository
-	if isGitRepo() && tracker != nil {
-		if err := tracker.StageAllFiles(verbose); err != nil {
-			return fmt.Errorf("failed to stage workflow files: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// CompileWorkflowWithValidation compiles a workflow with always-on YAML validation for CLI usage
-func CompileWorkflowWithValidation(compiler *workflow.Compiler, filePath string, verbose bool) error {
-	// Compile the workflow first
-	if err := compiler.CompileWorkflow(filePath); err != nil {
-		return err
-	}
-
-	// Always validate that the generated lock file is valid YAML (CLI requirement)
-	lockFile := strings.TrimSuffix(filePath, ".md") + ".lock.yml"
-	if _, err := os.Stat(lockFile); err != nil {
-		// Lock file doesn't exist (likely due to no-emit), skip YAML validation
-		return nil
-	}
-
-	if verbose {
-		fmt.Fprintln(os.Stderr, console.FormatInfoMessage("Validating generated lock file YAML syntax..."))
-	}
-
-	lockContent, err := os.ReadFile(lockFile)
-	if err != nil {
-		return fmt.Errorf("failed to read generated lock file for validation: %w", err)
-	}
-
-	// Validate the lock file is valid YAML
-	var yamlValidationTest any
-	if err := yaml.Unmarshal(lockContent, &yamlValidationTest); err != nil {
-		return fmt.Errorf("generated lock file is not valid YAML: %w", err)
-	}
-
-	return nil
-}
-
-// CompileConfig holds configuration options for compiling workflows
-type CompileConfig struct {
-	MarkdownFiles       []string // Files to compile (empty for all files)
-	Verbose             bool     // Enable verbose output
-	EngineOverride      string   // Override AI engine setting
-	Validate            bool     // Enable schema validation
-	Watch               bool     // Enable watch mode
-	WorkflowDir         string   // Custom workflow directory
-	SkipInstructions    bool     // Skip instruction validation
-	NoEmit              bool     // Validate without generating lock files
-	Purge               bool     // Remove orphaned lock files
-	TrialMode           bool     // Enable trial mode (suppress safe outputs)
-	TrialTargetRepoSlug string   // Target repository for trial mode
-	Strict              bool     // Enable strict mode validation
-}
-
-func CompileWorkflows(config CompileConfig) ([]*workflow.WorkflowData, error) {
-	markdownFiles := config.MarkdownFiles
-	verbose := config.Verbose
-	engineOverride := config.EngineOverride
-	validate := config.Validate
-	watch := config.Watch
-	workflowDir := config.WorkflowDir
-	skipInstructions := config.SkipInstructions
-	noEmit := config.NoEmit
-	purge := config.Purge
-	trialMode := config.TrialMode
-	trialTargetRepoSlug := config.TrialTargetRepoSlug
-	strict := config.Strict
-	// Validate purge flag usage
-	if purge && len(markdownFiles) > 0 {
-		return nil, fmt.Errorf("--purge flag can only be used when compiling all markdown files (no specific files specified)")
-	}
-
-	// Validate and set default for workflow directory
-	if workflowDir == "" {
-		workflowDir = ".github/workflows"
-	} else {
-		// Ensure the path is relative
-		if filepath.IsAbs(workflowDir) {
-			return nil, fmt.Errorf("workflows-dir must be a relative path, got: %s", workflowDir)
-		}
-		// Clean the path to avoid issues with ".." or other problematic elements
-		workflowDir = filepath.Clean(workflowDir)
-	}
-
-	// Create compiler with verbose flag and AI engine override
-	compiler := workflow.NewCompiler(verbose, engineOverride, GetVersion())
-
-	// Set validation based on the validate flag (false by default for compatibility)
-	compiler.SetSkipValidation(!validate)
-
-	// Set noEmit flag to validate without generating lock files
-	compiler.SetNoEmit(noEmit)
-
-	// Set strict mode if specified
-	compiler.SetStrictMode(strict)
-
-	// Set trial mode if specified
-	if trialMode {
-		compiler.SetTrialMode(true)
-		if trialTargetRepoSlug != "" {
-			compiler.SetTrialTargetRepo(trialTargetRepoSlug)
-		}
-	}
-
-	if watch {
-		// Watch mode: watch for file changes and recompile automatically
-		// For watch mode, we only support a single file for now
-		var markdownFile string
-		if len(markdownFiles) > 0 {
-			if len(markdownFiles) > 1 {
-				fmt.Fprintln(os.Stderr, console.FormatWarningMessage("Watch mode only supports a single file, using the first one"))
-			}
-			// Resolve the workflow file to get the full path
-			resolvedFile, err := resolveWorkflowFile(markdownFiles[0], verbose)
-			if err != nil {
-				return nil, fmt.Errorf("failed to resolve workflow '%s': %w", markdownFiles[0], err)
-			}
-			markdownFile = resolvedFile
-		}
-		return nil, watchAndCompileWorkflows(markdownFile, compiler, verbose)
-	}
-
-	var workflowDataList []*workflow.WorkflowData
-
-	if len(markdownFiles) > 0 {
-		// Compile specific workflow files
-		var compiledCount int
-		for _, markdownFile := range markdownFiles {
-			// Resolve workflow ID or file path to actual file path
-			resolvedFile, err := resolveWorkflowFile(markdownFile, verbose)
-			if err != nil {
-				return nil, fmt.Errorf("failed to resolve workflow '%s': %w", markdownFile, err)
-			}
-
-			// Parse workflow file to get data
-			workflowData, err := compiler.ParseWorkflowFile(resolvedFile)
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse workflow file %s: %w", resolvedFile, err)
-			}
-			workflowDataList = append(workflowDataList, workflowData)
-
-			if verbose {
-				fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Compiling %s", resolvedFile)))
-			}
-			if err := CompileWorkflowWithValidation(compiler, resolvedFile, verbose); err != nil {
-				// Always put error on a new line and don't wrap with "failed to compile workflow"
-				fmt.Fprintln(os.Stderr, err.Error())
-				return nil, fmt.Errorf("compilation failed")
-			}
-			compiledCount++
-		}
-
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Successfully compiled %d workflow file(s)", compiledCount)))
-		}
-
-		// Ensure .gitattributes marks .lock.yml files as generated
-		if err := ensureGitAttributes(); err != nil {
-			if verbose {
-				fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to update .gitattributes: %v", err)))
-			}
-		} else if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatSuccessMessage("Updated .gitattributes to mark .lock.yml files as generated"))
-		}
-
-		// Ensure copilot instructions are present
-		if err := ensureCopilotInstructions(verbose, skipInstructions); err != nil {
-			if verbose {
-				fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to update copilot instructions: %v", err)))
-			}
-		}
-
-		return workflowDataList, nil
-	}
-
-	// Find git root for consistent behavior
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return nil, fmt.Errorf("compile without arguments requires being in a git repository: %w", err)
-	}
-
-	// Compile all markdown files in the specified workflow directory relative to git root
-	workflowsDir := filepath.Join(gitRoot, workflowDir)
-	if _, err := os.Stat(workflowsDir); os.IsNotExist(err) {
-		return nil, fmt.Errorf("the %s directory does not exist in git root (%s)", workflowDir, gitRoot)
-	}
-
-	if verbose {
-		fmt.Printf("Scanning for markdown files in %s\n", workflowsDir)
-	}
-
-	// Find all markdown files
-	mdFiles, err := filepath.Glob(filepath.Join(workflowsDir, "*.md"))
-	if err != nil {
-		return nil, fmt.Errorf("failed to find markdown files: %w", err)
-	}
-
-	if len(mdFiles) == 0 {
-		return nil, fmt.Errorf("no markdown files found in %s", workflowsDir)
-	}
-
-	if verbose {
-		fmt.Printf("Found %d markdown files to compile\n", len(mdFiles))
-	}
-
-	// Handle purge logic: collect existing .lock.yml files before compilation
-	var existingLockFiles []string
-	var expectedLockFiles []string
-	if purge {
-		// Find all existing .lock.yml files
-		existingLockFiles, err = filepath.Glob(filepath.Join(workflowsDir, "*.lock.yml"))
-		if err != nil {
-			return nil, fmt.Errorf("failed to find existing lock files: %w", err)
-		}
-
-		// Create expected lock files list based on markdown files
-		for _, mdFile := range mdFiles {
-			lockFile := strings.TrimSuffix(mdFile, ".md") + ".lock.yml"
-			expectedLockFiles = append(expectedLockFiles, lockFile)
-		}
-
-		if verbose && len(existingLockFiles) > 0 {
-			fmt.Fprintln(os.Stderr, console.FormatInfoMessage(fmt.Sprintf("Found %d existing .lock.yml files", len(existingLockFiles))))
-		}
-	}
-
-	// Compile each file
-	for _, file := range mdFiles {
-		// Parse workflow file to get data
-		workflowData, err := compiler.ParseWorkflowFile(file)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse workflow file %s: %w", file, err)
-		}
-		workflowDataList = append(workflowDataList, workflowData)
-
-		if err := CompileWorkflowWithValidation(compiler, file, verbose); err != nil {
-			return nil, err
-		}
-	}
-
-	if verbose {
-		fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Successfully compiled all %d workflow files", len(mdFiles))))
-	}
-
-	// Handle purge logic: delete orphaned .lock.yml files
-	if purge && len(existingLockFiles) > 0 {
-		// Find lock files that should be deleted (exist but aren't expected)
-		expectedLockFileSet := make(map[string]bool)
-		for _, expected := range expectedLockFiles {
-			expectedLockFileSet[expected] = true
-		}
-
-		var orphanedFiles []string
-		for _, existing := range existingLockFiles {
-			if !expectedLockFileSet[existing] {
-				orphanedFiles = append(orphanedFiles, existing)
-			}
-		}
-
-		// Delete orphaned lock files
-		if len(orphanedFiles) > 0 {
-			for _, orphanedFile := range orphanedFiles {
-				if err := os.Remove(orphanedFile); err != nil {
-					fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to remove orphaned lock file %s: %v", filepath.Base(orphanedFile), err)))
-				} else {
-					fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Removed orphaned lock file: %s", filepath.Base(orphanedFile))))
-				}
-			}
-			if verbose {
-				fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Purged %d orphaned .lock.yml files", len(orphanedFiles))))
-			}
-		} else if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatInfoMessage("No orphaned .lock.yml files found to purge"))
-		}
-	}
-
-	// Ensure .gitattributes marks .lock.yml files as generated
-	if err := ensureGitAttributes(); err != nil {
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to update .gitattributes: %v", err)))
-		}
-	} else if verbose {
-		fmt.Fprintln(os.Stderr, console.FormatSuccessMessage("Updated .gitattributes to mark .lock.yml files as generated"))
-	}
-
-	// Ensure copilot instructions are present
-	if err := ensureCopilotInstructions(verbose, skipInstructions); err != nil {
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to update copilot instructions: %v", err)))
-		}
-	}
-
-	// Ensure agentic workflow prompt is present
-	if err := ensureAgenticWorkflowPrompt(verbose, skipInstructions); err != nil {
-		if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Failed to update agentic workflow prompt: %v", err)))
-		}
-	}
-
-	return workflowDataList, nil
-}
-
-// watchAndCompileWorkflows watches for changes to workflow files and recompiles them automatically
-func watchAndCompileWorkflows(markdownFile string, compiler *workflow.Compiler, verbose bool) error {
-	// Find git root for consistent behavior
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return fmt.Errorf("watch mode requires being in a git repository: %w", err)
-	}
-
-	workflowsDir := filepath.Join(gitRoot, ".github/workflows")
-	if _, err := os.Stat(workflowsDir); os.IsNotExist(err) {
-		return fmt.Errorf("the .github/workflows directory does not exist in git root (%s)", gitRoot)
-	}
-
-	// If a specific file is provided, watch only that file and its directory
-	if markdownFile != "" {
-		if !filepath.IsAbs(markdownFile) {
-			markdownFile = filepath.Join(workflowsDir, markdownFile)
-		}
-		if _, err := os.Stat(markdownFile); os.IsNotExist(err) {
-			return fmt.Errorf("specified markdown file does not exist: %s", markdownFile)
-		}
-	}
-
-	// Set up file system watcher
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		return fmt.Errorf("failed to create file watcher: %w", err)
-	}
-	defer watcher.Close()
-
-	// Add the workflows directory to the watcher
-	if err := watcher.Add(workflowsDir); err != nil {
-		return fmt.Errorf("failed to watch directory %s: %w", workflowsDir, err)
-	}
-
-	// Also watch subdirectories for include files (recursive watching)
-	err = filepath.Walk(workflowsDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return nil // Skip errors but continue walking
-		}
-		if info.IsDir() && path != workflowsDir {
-			// Add subdirectories to the watcher
-			if err := watcher.Add(path); err != nil {
-				if verbose {
-					fmt.Printf("Warning: Failed to watch subdirectory %s: %v\n", path, err)
-				}
-			} else if verbose {
-				fmt.Printf("Watching subdirectory: %s\n", path)
-			}
-		}
-		return nil
-	})
-	if err != nil && verbose {
-		fmt.Printf("Warning: Failed to walk subdirectories: %v\n", err)
-	}
-
-	// Always emit the begin pattern for task integration
-	if markdownFile != "" {
-		fmt.Printf("Watching for file changes to %s...\n", markdownFile)
-	} else {
-		fmt.Printf("Watching for file changes in %s...\n", workflowsDir)
-	}
-
-	if verbose {
-		fmt.Fprintln(os.Stderr, "Press Ctrl+C to stop watching.")
-	}
-
-	// Set up signal handling for graceful shutdown
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
-
-	// Debouncing setup
-	const debounceDelay = 300 * time.Millisecond
-	var debounceTimer *time.Timer
-	modifiedFiles := make(map[string]struct{})
-
-	// Compile initially if no specific file provided
-	if markdownFile == "" {
-		fmt.Fprintln(os.Stderr, "Watching for file changes")
-		if verbose {
-			fmt.Fprintln(os.Stderr, "🔨 Initial compilation of all workflow files...")
-		}
-		if err := compileAllWorkflowFiles(compiler, workflowsDir, verbose); err != nil {
-			// Always show initial compilation errors, not just in verbose mode
-			fmt.Fprintln(os.Stderr, console.FormatWarningMessage(fmt.Sprintf("Initial compilation failed: %v", err)))
-		}
-		fmt.Fprintln(os.Stderr, "Recompiled")
-	} else {
-		fmt.Fprintln(os.Stderr, "Watching for file changes")
-		if verbose {
-			fmt.Fprintf(os.Stderr, "🔨 Initial compilation of %s...\n", markdownFile)
-		}
-		if err := CompileWorkflowWithValidation(compiler, markdownFile, verbose); err != nil {
-			// Always show initial compilation errors on new line without wrapping
-			fmt.Fprintln(os.Stderr, err.Error())
-		}
-		fmt.Fprintln(os.Stderr, "Recompiled")
-	}
-
-	// Main watch loop
-	for {
-		select {
-		case event, ok := <-watcher.Events:
-			if !ok {
-				return fmt.Errorf("watcher channel closed")
-			}
-
-			// Only process markdown files and ignore lock files
-			if !strings.HasSuffix(event.Name, ".md") {
-				continue
-			}
-
-			// If watching a specific file, only process that file
-			if markdownFile != "" && event.Name != markdownFile {
-				continue
-			}
-
-			if verbose {
-				fmt.Printf("📝 Detected change: %s (%s)\n", event.Name, event.Op.String())
-			}
-
-			// Handle file operations
-			switch {
-			case event.Has(fsnotify.Remove):
-				// Handle file deletion
-				handleFileDeleted(event.Name, verbose)
-			case event.Has(fsnotify.Write) || event.Has(fsnotify.Create):
-				// Handle file modification or creation - add to debounced compilation
-				modifiedFiles[event.Name] = struct{}{}
-
-				// Reset debounce timer
-				if debounceTimer != nil {
-					debounceTimer.Stop()
-				}
-				debounceTimer = time.AfterFunc(debounceDelay, func() {
-					filesToCompile := make([]string, 0, len(modifiedFiles))
-					for file := range modifiedFiles {
-						filesToCompile = append(filesToCompile, file)
-					}
-					// Clear the modifiedFiles map
-					modifiedFiles = make(map[string]struct{})
-
-					// Compile the modified files
-					compileModifiedFiles(compiler, filesToCompile, verbose)
-				})
-			}
-
-		case err, ok := <-watcher.Errors:
-			if !ok {
-				return fmt.Errorf("watcher error channel closed")
-			}
-			if verbose {
-				fmt.Printf("⚠️  Watcher error: %v\n", err)
-			}
-
-		case <-sigChan:
-			if verbose {
-				fmt.Fprintln(os.Stderr, "\n🛑 Stopping watch mode...")
-			}
-			if debounceTimer != nil {
-				debounceTimer.Stop()
-			}
-			return nil
-		}
-	}
-}
-
-// compileAllWorkflowFiles compiles all markdown files in the workflows directory
-func compileAllWorkflowFiles(compiler *workflow.Compiler, workflowsDir string, verbose bool) error {
-	// Find all markdown files
-	mdFiles, err := filepath.Glob(filepath.Join(workflowsDir, "*.md"))
-	if err != nil {
-		return fmt.Errorf("failed to find markdown files: %w", err)
-	}
-
-	if len(mdFiles) == 0 {
-		if verbose {
-			fmt.Printf("No markdown files found in %s\n", workflowsDir)
-		}
-		return nil
-	}
-
-	// Compile each file
-	for _, file := range mdFiles {
-		if verbose {
-			fmt.Printf("🔨 Compiling: %s\n", file)
-		}
-		if err := CompileWorkflowWithValidation(compiler, file, verbose); err != nil {
-			// Always show compilation errors on new line
-			fmt.Fprintln(os.Stderr, err.Error())
-		} else if verbose {
-			fmt.Println(console.FormatSuccessMessage(fmt.Sprintf("Compiled %s", file)))
-		}
-	}
-
-	// Ensure .gitattributes marks .lock.yml files as generated
-	if err := ensureGitAttributes(); err != nil {
-		if verbose {
-			fmt.Printf("⚠️  Failed to update .gitattributes: %v\n", err)
-		}
-	}
-
-	return nil
-}
-
-// compileModifiedFiles compiles a list of modified markdown files
-func compileModifiedFiles(compiler *workflow.Compiler, files []string, verbose bool) {
-	if len(files) == 0 {
-		return
-	}
-
-	fmt.Fprintln(os.Stderr, "Watching for file changes")
-	if verbose {
-		fmt.Fprintf(os.Stderr, "🔨 Compiling %d modified file(s)...\n", len(files))
-	}
-
-	for _, file := range files {
-		// Check if file still exists (might have been deleted between detection and compilation)
-		if _, err := os.Stat(file); os.IsNotExist(err) {
-			if verbose {
-				fmt.Printf("📝 File %s was deleted, skipping compilation\n", file)
-			}
-			continue
-		}
-
-		if verbose {
-			fmt.Fprintf(os.Stderr, "🔨 Compiling: %s\n", file)
-		}
-
-		if err := CompileWorkflowWithValidation(compiler, file, verbose); err != nil {
-			// Always show compilation errors on new line
-			fmt.Fprintln(os.Stderr, err.Error())
-		} else if verbose {
-			fmt.Fprintln(os.Stderr, console.FormatSuccessMessage(fmt.Sprintf("Compiled %s", file)))
-		}
-	}
-
-	// Ensure .gitattributes marks .lock.yml files as generated
-	if err := ensureGitAttributes(); err != nil {
-		if verbose {
-			fmt.Printf("⚠️  Failed to update .gitattributes: %v\n", err)
-		}
-	}
-
-	fmt.Println("Recompiled")
-}
-
-// handleFileDeleted handles the deletion of a markdown file by removing its corresponding lock file
-func handleFileDeleted(mdFile string, verbose bool) {
-	// Generate the corresponding lock file path
-	lockFile := strings.TrimSuffix(mdFile, ".md") + ".lock.yml"
-
-	// Check if the lock file exists and remove it
-	if _, err := os.Stat(lockFile); err == nil {
-		if err := os.Remove(lockFile); err != nil {
-			if verbose {
-				fmt.Printf("⚠️  Failed to remove lock file %s: %v\n", lockFile, err)
-			}
-		} else {
-			if verbose {
-				fmt.Printf("🗑️  Removed corresponding lock file: %s\n", lockFile)
-			}
-		}
-	}
-}
-
-// RemoveWorkflows removes workflows matching a pattern
-func RemoveWorkflows(pattern string, keepOrphans bool) error {
-	workflowsDir := getWorkflowsDir()
-
-	if _, err := os.Stat(workflowsDir); os.IsNotExist(err) {
-		fmt.Println("No .github/workflows directory found.")
-		return nil
-	}
-
-	// Find all markdown files in .github/workflows
-	mdFiles, err := filepath.Glob(filepath.Join(workflowsDir, "*.md"))
-	if err != nil {
-		return fmt.Errorf("failed to find workflow files: %w", err)
-	}
-
-	if len(mdFiles) == 0 {
-		fmt.Println("No workflow files found to remove.")
-		return nil
-	}
-
-	var filesToRemove []string
-
-	// If no pattern specified, list all files for user to see
-	if pattern == "" {
-		fmt.Println("Available workflows to remove:")
-		for _, file := range mdFiles {
-			workflowName, _ := extractWorkflowNameFromFile(file)
-			base := filepath.Base(file)
-			name := strings.TrimSuffix(base, ".md")
-			if workflowName != "" {
-				fmt.Printf("  %-20s - %s\n", name, workflowName)
-			} else {
-				fmt.Printf("  %s\n", name)
-			}
-		}
-		fmt.Println("\nUsage: " + constants.CLIExtensionPrefix + " remove <pattern>")
-		return nil
-	}
-
-	// Find matching files by workflow name or filename
-	for _, file := range mdFiles {
-		base := filepath.Base(file)
-		filename := strings.TrimSuffix(base, ".md")
-		workflowName, _ := extractWorkflowNameFromFile(file)
-
-		// Check if pattern matches filename or workflow name
-		if strings.Contains(strings.ToLower(filename), strings.ToLower(pattern)) ||
-			strings.Contains(strings.ToLower(workflowName), strings.ToLower(pattern)) {
-			filesToRemove = append(filesToRemove, file)
-		}
-	}
-
-	if len(filesToRemove) == 0 {
-		fmt.Printf("No workflows found matching pattern: %s\n", pattern)
-		return nil
-	}
-
-	// Preview orphaned includes that would be removed (if orphan removal is enabled)
-	var orphanedIncludes []string
-	if !keepOrphans {
-		var err error
-		orphanedIncludes, err = previewOrphanedIncludes(filesToRemove, false)
-		if err != nil {
-			fmt.Printf("Warning: Failed to preview orphaned includes: %v\n", err)
-			orphanedIncludes = []string{} // Continue with empty list
-		}
-	}
-
-	// Show what will be removed
-	fmt.Printf("The following workflows will be removed:\n")
-	for _, file := range filesToRemove {
-		workflowName, _ := extractWorkflowNameFromFile(file)
-		if workflowName != "" {
-			fmt.Printf("  %s - %s\n", filepath.Base(file), workflowName)
-		} else {
-			fmt.Printf("  %s\n", filepath.Base(file))
-		}
-
-		// Also check for corresponding .lock.yml file in .github/workflows
-		lockFile := strings.TrimSuffix(file, ".md") + ".lock.yml"
-		if _, err := os.Stat(lockFile); err == nil {
-			fmt.Printf("  %s (compiled workflow)\n", filepath.Base(lockFile))
-		}
-	}
-
-	// Show orphaned includes that will also be removed
-	if len(orphanedIncludes) > 0 {
-		fmt.Printf("\nThe following orphaned include files will also be removed (suppress with --keep-orphans):\n")
-		for _, include := range orphanedIncludes {
-			fmt.Printf("  %s (orphaned include)\n", include)
-		}
-	}
-
-	// Ask for confirmation
-	fmt.Print("\nAre you sure you want to remove these workflows? [y/N]: ")
-	reader := bufio.NewReader(os.Stdin)
-	response, _ := reader.ReadString('\n')
-	response = strings.TrimSpace(strings.ToLower(response))
-
-	if response != "y" && response != "yes" {
-		fmt.Println("Operation cancelled.")
-		return nil
-	}
-
-	// Remove the files
-	var removedFiles []string
-	for _, file := range filesToRemove {
-		if err := os.Remove(file); err != nil {
-			fmt.Println(console.FormatWarningMessage(fmt.Sprintf("Failed to remove %s: %v", file, err)))
-		} else {
-			fmt.Println(console.FormatSuccessMessage(fmt.Sprintf("Removed: %s", filepath.Base(file))))
-			removedFiles = append(removedFiles, file)
-		}
-
-		// Also remove corresponding .lock.yml file
-		lockFile := strings.TrimSuffix(file, ".md") + ".lock.yml"
-		if _, err := os.Stat(lockFile); err == nil {
-			if err := os.Remove(lockFile); err != nil {
-				fmt.Printf("Warning: Failed to remove %s: %v\n", lockFile, err)
-			} else {
-				fmt.Printf("Removed: %s\n", filepath.Base(lockFile))
-			}
-		}
-	}
-
-	// Clean up orphaned include files (if orphan removal is enabled)
-	if len(removedFiles) > 0 && !keepOrphans {
-		if err := cleanupOrphanedIncludes(false); err != nil {
-			fmt.Printf("Warning: Failed to clean up orphaned includes: %v\n", err)
-		}
-	}
-
-	// Stage changes to git if in a git repository
-	if len(removedFiles) > 0 && isGitRepo() {
-		stageWorkflowChanges()
-	}
-
-	return nil
-}
-
-// StatusWorkflows shows status of workflows
-// getMarkdownWorkflowFiles finds all markdown files in .github/workflows directory
-func getMarkdownWorkflowFiles() ([]string, error) {
-	workflowsDir := getWorkflowsDir()
-
-	if _, err := os.Stat(workflowsDir); os.IsNotExist(err) {
-		return nil, fmt.Errorf("no .github/workflows directory found")
-	}
-
-	// Find all markdown files in .github/workflows
-	mdFiles, err := filepath.Glob(filepath.Join(workflowsDir, "*.md"))
-	if err != nil {
-		return nil, fmt.Errorf("failed to find workflow files: %w", err)
-	}
-
-	return mdFiles, nil
-}
-
-func StatusWorkflows(pattern string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Checking status of workflow files\n")
-		if pattern != "" {
-			fmt.Printf("Filtering by pattern: %s\n", pattern)
-		}
-	}
-
-	mdFiles, err := getMarkdownWorkflowFiles()
-	if err != nil {
-		fmt.Println(err.Error())
-		return nil
-	}
-
-	if len(mdFiles) == 0 {
-		fmt.Println("No workflow files found.")
-		return nil
-	}
-
-	if verbose {
-		fmt.Printf("Found %d markdown workflow files\n", len(mdFiles))
-		fmt.Printf("Fetching GitHub workflow status...\n")
-	}
-
-	// Get GitHub workflows data
-	githubWorkflows, err := fetchGitHubWorkflows(verbose)
-	if err != nil {
-		if verbose {
-			fmt.Printf("Verbose: Failed to fetch GitHub workflows: %v\n", err)
-		}
-		fmt.Printf("Warning: Could not fetch GitHub workflow status: %v\n", err)
-		githubWorkflows = make(map[string]*GitHubWorkflow)
-	} else if verbose {
-		fmt.Printf("Successfully fetched %d GitHub workflows\n", len(githubWorkflows))
-	}
-
-	// Build table configuration
-	headers := []string{"Name", "Installed", "Up-to-date", "Status", "Time Remaining"}
-	var rows [][]string
-
-	for _, file := range mdFiles {
-		base := filepath.Base(file)
-		name := strings.TrimSuffix(base, ".md")
-
-		// Skip if pattern specified and doesn't match
-		if pattern != "" && !strings.Contains(strings.ToLower(name), strings.ToLower(pattern)) {
-			continue
-		}
-
-		// Check if compiled (.lock.yml file is in .github/workflows)
-		lockFile := strings.TrimSuffix(file, ".md") + ".lock.yml"
-		compiled := "No"
-		upToDate := "N/A"
-		timeRemaining := "N/A"
-
-		if _, err := os.Stat(lockFile); err == nil {
-			compiled = "Yes"
-
-			// Check if up to date
-			mdStat, _ := os.Stat(file)
-			lockStat, _ := os.Stat(lockFile)
-			if mdStat.ModTime().After(lockStat.ModTime()) {
-				upToDate = "No"
-			} else {
-				upToDate = "Yes"
-			}
-
-			// Extract stop-time from lock file
-			if stopTime := extractStopTimeFromLockFile(lockFile); stopTime != "" {
-				timeRemaining = calculateTimeRemaining(stopTime)
-			}
-		}
-
-		// Get GitHub workflow status
-		status := "Unknown"
-		if workflow, exists := githubWorkflows[name]; exists {
-			if workflow.State == "disabled_manually" {
-				status = "disabled"
-			} else {
-				status = workflow.State
-			}
-		}
-
-		// Build row data
-		row := []string{name, compiled, upToDate, status, timeRemaining}
-		rows = append(rows, row)
-	}
-
-	// Render the table
-	tableConfig := console.TableConfig{
-		Title:   "Workflow Status",
-		Headers: headers,
-		Rows:    rows,
-	}
-	fmt.Print(console.RenderTable(tableConfig))
-
-	return nil
-}
-
-// extractStopTimeFromLockFile extracts the STOP_TIME value from a compiled workflow lock file
-func extractStopTimeFromLockFile(lockFilePath string) string {
-	content, err := os.ReadFile(lockFilePath)
-	if err != nil {
-		return ""
-	}
-
-	// Look for the STOP_TIME line in the safety checks section
-	// Pattern: STOP_TIME="YYYY-MM-DD HH:MM:SS"
-	lines := strings.Split(string(content), "\n")
-	for _, line := range lines {
-		if strings.Contains(line, "STOP_TIME=") {
-			// Extract the value between quotes
-			start := strings.Index(line, `"`) + 1
-			end := strings.LastIndex(line, `"`)
-			if start > 0 && end > start {
-				return line[start:end]
-			}
-		}
-	}
-	return ""
-}
-
-// calculateTimeRemaining calculates and formats the time remaining until stop-time
-func calculateTimeRemaining(stopTimeStr string) string {
-	if stopTimeStr == "" {
-		return "N/A"
-	}
-
-	// Parse the stop time in local timezone
-	stopTime, err := time.ParseInLocation("2006-01-02 15:04:05", stopTimeStr, time.Local)
-	if err != nil {
-		return "Invalid"
-	}
-
-	now := time.Now()
-	remaining := stopTime.Sub(now)
-
-	// If already past the stop time
-	if remaining <= 0 {
-		return "Expired"
-	}
-
-	// Format the remaining time in a human-readable way
-	days := int(remaining.Hours() / 24)
-	hours := int(remaining.Hours()) % 24
-	minutes := int(remaining.Minutes()) % 60
-
-	if days > 0 {
-		if days == 1 {
-			return fmt.Sprintf("%dd %dh", days, hours)
-		}
-		return fmt.Sprintf("%dd %dh", days, hours)
-	} else if hours > 0 {
-		return fmt.Sprintf("%dh %dm", hours, minutes)
-	} else if minutes > 0 {
-		return fmt.Sprintf("%dm", minutes)
-	} else {
-		return "< 1m"
-	}
-}
-
-// Helper functions
-
-func extractWorkflowNameFromFile(filePath string) (string, error) {
-	content, err := os.ReadFile(filePath)
-	if err != nil {
-		return "", err
-	}
-
-	// Extract markdown content (excluding frontmatter)
-	result, err := parser.ExtractFrontmatterFromContent(string(content))
-	if err != nil {
-		return "", err
-	}
-
-	// Look for first H1 header
-	scanner := bufio.NewScanner(strings.NewReader(result.Markdown))
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if strings.HasPrefix(line, "# ") {
-			return strings.TrimSpace(line[2:]), nil
-		}
-	}
-
-	// No H1 header found, generate default name from filename
-	baseName := filepath.Base(filePath)
-	baseName = strings.TrimSuffix(baseName, filepath.Ext(baseName))
-	baseName = strings.ReplaceAll(baseName, "-", " ")
-
-	// Capitalize first letter of each word
-	words := strings.Fields(baseName)
-	for i, word := range words {
-		if len(word) > 0 {
-			words[i] = strings.ToUpper(word[:1]) + word[1:]
-		}
-	}
-
-	return strings.Join(words, " "), nil
-}
-
-func updateWorkflowTitle(content string, number int) string {
-	// Find and update the first H1 header
-	lines := strings.Split(content, "\n")
-	for i, line := range lines {
-		if strings.HasPrefix(strings.TrimSpace(line), "# ") {
-			// Extract the title part and add number
-			title := strings.TrimSpace(line[2:])
-			lines[i] = fmt.Sprintf("# %s %d", title, number)
-			break
-		}
-	}
-	return strings.Join(lines, "\n")
-}
-
-func compileWorkflow(filePath string, verbose bool, engineOverride string) error {
-	// Create compiler and compile the workflow
-	compiler := workflow.NewCompiler(verbose, engineOverride, GetVersion())
-	if err := CompileWorkflowWithValidation(compiler, filePath, verbose); err != nil {
-		return err
-	}
-
-	// Ensure .gitattributes marks .lock.yml files as generated
-	if err := ensureGitAttributes(); err != nil {
-		if verbose {
-			fmt.Printf("Warning: Failed to update .gitattributes: %v\n", err)
-		}
-	}
-
-	// Note: Instructions are only written when explicitly requested via the compile command flag
-	// This helper function is used in contexts where instructions should not be automatically written
-
-	return nil
-}
-
-// compileWorkflowWithTracking compiles a workflow and tracks generated files
-func compileWorkflowWithTracking(filePath string, verbose bool, engineOverride string, tracker *FileTracker) error {
-	// Generate the expected lock file path
-	lockFile := strings.TrimSuffix(filePath, ".md") + ".lock.yml"
-
-	// Check if lock file exists before compilation
-	lockFileExists := false
-	if _, err := os.Stat(lockFile); err == nil {
-		lockFileExists = true
-	}
-
-	// Check if .gitattributes exists before ensuring it
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return err
-	}
-	gitAttributesPath := filepath.Join(gitRoot, ".gitattributes")
-	gitAttributesExists := false
-	if _, err := os.Stat(gitAttributesPath); err == nil {
-		gitAttributesExists = true
-	}
-
-	// Track the lock file before compilation
-	if lockFileExists {
-		tracker.TrackModified(lockFile)
-	} else {
-		tracker.TrackCreated(lockFile)
-	}
-
-	// Track .gitattributes file before modification
-	if gitAttributesExists {
-		tracker.TrackModified(gitAttributesPath)
-	} else {
-		tracker.TrackCreated(gitAttributesPath)
-	}
-
-	// Create compiler and set the file tracker
-	compiler := workflow.NewCompiler(verbose, engineOverride, GetVersion())
-	compiler.SetFileTracker(tracker)
-	if err := CompileWorkflowWithValidation(compiler, filePath, verbose); err != nil {
-		return err
-	}
-
-	// Ensure .gitattributes marks .lock.yml files as generated
-	if err := ensureGitAttributes(); err != nil {
-		if verbose {
-			fmt.Printf("Warning: Failed to update .gitattributes: %v\n", err)
-		}
-	}
-
-	return nil
-}
-
-// ensureCopilotInstructions ensures that .github/instructions/github-agentic-workflows.md contains the copilot instructions
-func ensureCopilotInstructions(verbose bool, skipInstructions bool) error {
-	if skipInstructions {
-		return nil // Skip writing instructions if flag is set
-	}
-
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return err // Not in a git repository, skip
-	}
-
-	copilotDir := filepath.Join(gitRoot, ".github", "instructions")
-	copilotInstructionsPath := filepath.Join(copilotDir, "github-agentic-workflows.instructions.md")
-
-	// Ensure the .github/instructions directory exists
-	if err := os.MkdirAll(copilotDir, 0755); err != nil {
-		return fmt.Errorf("failed to create .github/instructions directory: %w", err)
-	}
-
-	// Check if the instructions file already exists and matches the template
-	existingContent := ""
-	if content, err := os.ReadFile(copilotInstructionsPath); err == nil {
-		existingContent = string(content)
-	}
-
-	// Check if content matches our expected template
-	expectedContent := strings.TrimSpace(copilotInstructionsTemplate)
-	if strings.TrimSpace(existingContent) == expectedContent {
-		if verbose {
-			fmt.Printf("Copilot instructions are up-to-date: %s\n", copilotInstructionsPath)
-		}
-		return nil
-	}
-
-	// Write the copilot instructions file
-	if err := os.WriteFile(copilotInstructionsPath, []byte(copilotInstructionsTemplate), 0644); err != nil {
-		return fmt.Errorf("failed to write copilot instructions: %w", err)
-	}
-
-	if verbose {
-		if existingContent == "" {
-			fmt.Printf("Created copilot instructions: %s\n", copilotInstructionsPath)
-		} else {
-			fmt.Printf("Updated copilot instructions: %s\n", copilotInstructionsPath)
-		}
-	}
-
-	return nil
-}
-
-// ensureAgenticWorkflowPrompt ensures that .github/prompts/create-agentic-workflow.prompt.md contains the agentic workflow creation prompt
-func ensureAgenticWorkflowPrompt(verbose bool, skipInstructions bool) error {
-	if skipInstructions {
-		return nil // Skip writing prompt if flag is set
-	}
-
-	gitRoot, err := findGitRoot()
-	if err != nil {
-		return err // Not in a git repository, skip
-	}
-
-	promptsDir := filepath.Join(gitRoot, ".github", "prompts")
-	agenticWorkflowPromptPath := filepath.Join(promptsDir, "create-agentic-workflow.prompt.md")
-
-	// Ensure the .github/prompts directory exists
-	if err := os.MkdirAll(promptsDir, 0755); err != nil {
-		return fmt.Errorf("failed to create .github/prompts directory: %w", err)
-	}
-
-	// Check if the prompt file already exists and matches the template
-	existingContent := ""
-	if content, err := os.ReadFile(agenticWorkflowPromptPath); err == nil {
-		existingContent = string(content)
-	}
-
-	// Check if content matches our expected template
-	expectedContent := strings.TrimSpace(agenticWorkflowPromptTemplate)
-	if strings.TrimSpace(existingContent) == expectedContent {
-		if verbose {
-			fmt.Printf("Agentic workflow prompt is up-to-date: %s\n", agenticWorkflowPromptPath)
-		}
-		return nil
-	}
-
-	// Write the agentic workflow prompt file
-	if err := os.WriteFile(agenticWorkflowPromptPath, []byte(agenticWorkflowPromptTemplate), 0644); err != nil {
-		return fmt.Errorf("failed to write agentic workflow prompt: %w", err)
-	}
-
-	if verbose {
-		if existingContent == "" {
-			fmt.Printf("Created agentic workflow prompt: %s\n", agenticWorkflowPromptPath)
-		} else {
-			fmt.Printf("Updated agentic workflow prompt: %s\n", agenticWorkflowPromptPath)
-		}
-	}
-
-	return nil
-}
-
 func isGHCLIAvailable() bool {
 	cmd := exec.Command("gh", "--version")
 	return cmd.Run() == nil
-}
-
-// IncludeDependency represents a file dependency from @include directives
-type IncludeDependency struct {
-	SourcePath string // Path in the source (local)
-	TargetPath string // Relative path where it should be copied in .github/workflows
-	IsOptional bool   // Whether this is an optional include (@include?)
-}
-
-// collectIncludeDependencies recursively collects all @include dependencies from a workflow file
-func collectIncludeDependencies(content, workflowPath, workflowsDir string) ([]IncludeDependency, error) {
-	var dependencies []IncludeDependency
-	seen := make(map[string]bool) // Track already processed files to avoid cycles
-
-	// Get the directory of the workflow file for resolving relative paths
-	var workflowDir = filepath.Dir(workflowPath)
-
-	err := collectIncludesRecursive(content, workflowDir, workflowsDir, &dependencies, seen)
-	return dependencies, err
-}
-
-// collectIncludesRecursive recursively processes @include directives in content
-func collectIncludesRecursive(content, baseDir, workflowsDir string, dependencies *[]IncludeDependency, seen map[string]bool) error {
-	includePattern := regexp.MustCompile(`^@include(\?)?\s+(.+)$`)
-
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	for scanner.Scan() {
-		line := scanner.Text()
-		if matches := includePattern.FindStringSubmatch(line); matches != nil {
-			isOptional := matches[1] == "?"
-			includePath := strings.TrimSpace(matches[2])
-
-			// Handle section references (file.md#Section)
-			var filePath string
-			if strings.Contains(includePath, "#") {
-				parts := strings.SplitN(includePath, "#", 2)
-				filePath = parts[0]
-			} else {
-				filePath = includePath
-			}
-
-			// Resolve the full source path
-			var fullSourcePath = filepath.Join(baseDir, filePath)
-
-			// Skip if we've already processed this file
-			if seen[fullSourcePath] {
-				continue
-			}
-			seen[fullSourcePath] = true
-
-			// Add dependency (even for optional includes that might not exist yet)
-			dep := IncludeDependency{
-				SourcePath: fullSourcePath,
-				TargetPath: filePath, // Keep relative path for target
-				IsOptional: isOptional,
-			}
-			*dependencies = append(*dependencies, dep)
-
-			// Read the included file and process its includes recursively
-			var includedContent []byte
-			var err error
-			includedContent, err = os.ReadFile(fullSourcePath)
-
-			if err != nil {
-				// If we can't read the file, add it anyway but don't recurse
-				continue
-			}
-
-			// Extract markdown content from the included file
-			markdownContent, err := parser.ExtractMarkdownContent(string(includedContent))
-			if err != nil {
-				continue // Skip if we can't extract markdown
-			}
-
-			// Recursively process includes in the included file
-			includedDir := filepath.Dir(fullSourcePath)
-			if err := collectIncludesRecursive(markdownContent, includedDir, workflowsDir, dependencies, seen); err != nil {
-				// Log error but continue processing other includes
-				fmt.Printf("Warning: Error processing includes in %s: %v\n", fullSourcePath, err)
-			}
-		}
-	}
-
-	return scanner.Err()
-}
-
-// Package represents an installed package
-type Package struct {
-	Name      string
-	Path      string
-	Workflows []string
-	CommitSHA string
-}
-
-// parseRepoSpec parses repository specification like "org/repo@version" or "org/repo@branch" or "org/repo@commit"
-func parseRepoSpec(repoSpec string) (repo, version string, err error) {
-	parts := strings.SplitN(repoSpec, "@", 2)
-	repo = parts[0]
-
-	// Validate repository format (org/repo)
-	repoParts := strings.Split(repo, "/")
-	if len(repoParts) != 2 || repoParts[0] == "" || repoParts[1] == "" {
-		return "", "", fmt.Errorf("repository must be in format 'org/repo'")
-	}
-
-	if len(parts) == 2 {
-		version = parts[1]
-	}
-
-	return repo, version, nil
-}
-
-// isCommitSHA checks if a version string looks like a commit SHA (40-character hex string)
-func isCommitSHA(version string) bool {
-	if len(version) != 40 {
-		return false
-	}
-	// Check if all characters are hexadecimal
-	for _, char := range version {
-		if !((char >= '0' && char <= '9') || (char >= 'a' && char <= 'f') || (char >= 'A' && char <= 'F')) {
-			return false
-		}
-	}
-	return true
-}
-
-// WorkflowSpec represents a parsed workflow specification
-type WorkflowSpec struct {
-	Repo         string // e.g., "owner/repo"
-	WorkflowPath string // e.g., "workflows/workflow-name" or "workflow-name"
-	Version      string // optional version/tag/SHA
-}
-
-// isValidGitHubIdentifier checks if a string looks like a valid GitHub username or repository name
-// GitHub allows alphanumeric characters, hyphens, and underscores, but cannot start or end with hyphen
-func isValidGitHubIdentifier(identifier string) bool {
-	if len(identifier) == 0 {
-		return false
-	}
-
-	// Cannot start or end with hyphen
-	if identifier[0] == '-' || identifier[len(identifier)-1] == '-' {
-		return false
-	}
-
-	// Must contain only alphanumeric chars, hyphens, and underscores
-	for _, char := range identifier {
-		if !((char >= 'a' && char <= 'z') ||
-			(char >= 'A' && char <= 'Z') ||
-			(char >= '0' && char <= '9') ||
-			char == '-' || char == '_') {
-			return false
-		}
-	}
-
-	return true
-}
-
-// parseWorkflowSpec parses a workflow specification in the new format
-// Format: owner/repo/workflows/workflow-name[@version] or owner/repo/workflow-name[@version]
-func parseWorkflowSpec(spec string) (*WorkflowSpec, error) {
-	// Handle version first (anything after @)
-	parts := strings.SplitN(spec, "@", 2)
-	specWithoutVersion := parts[0]
-	var version string
-	if len(parts) == 2 {
-		version = parts[1]
-	}
-
-	// Split by slashes
-	slashParts := strings.Split(specWithoutVersion, "/")
-
-	// Must have at least 3 parts: owner/repo/workflow-path
-	if len(slashParts) < 3 {
-		return nil, fmt.Errorf("workflow specification must be in format 'owner/repo/workflow-name[@version]'")
-	}
-
-	owner := slashParts[0]
-	repo := slashParts[1]
-	workflowPath := strings.Join(slashParts[2:], "/")
-
-	// Validate owner and repo parts are not empty
-	if owner == "" || repo == "" {
-		return nil, fmt.Errorf("invalid workflow specification: owner and repo cannot be empty")
-	}
-
-	// Basic validation that owner and repo look like GitHub identifiers
-	if !isValidGitHubIdentifier(owner) || !isValidGitHubIdentifier(repo) {
-		return nil, fmt.Errorf("invalid workflow specification: '%s/%s' does not look like a valid GitHub repository", owner, repo)
-	}
-
-	// Handle different cases based on the number of path parts
-	if len(slashParts) == 3 {
-		// Three-part spec: owner/repo/workflow-name
-		// Files are stored at root level in packages (workflows/ prefix is only in the source repo)
-		if strings.HasSuffix(workflowPath, ".md") {
-			// Remove .md extension for internal processing
-			workflowPath = strings.TrimSuffix(workflowPath, ".md")
-		}
-		// No .md extension, use as-is - files are at root level in packages
-	} else {
-		// Four or more parts: owner/repo/workflows/workflow-name or owner/repo/path/to/workflow-name
-		// Require .md extension to be explicit
-		if !strings.HasSuffix(workflowPath, ".md") {
-			return nil, fmt.Errorf("workflow specification with path must end with '.md' extension: %s", workflowPath)
-		}
-		// Remove .md extension for internal processing
-		workflowPath = strings.TrimSuffix(workflowPath, ".md")
-	}
-
-	return &WorkflowSpec{
-		Repo:         fmt.Sprintf("%s/%s", owner, repo),
-		WorkflowPath: workflowPath,
-		Version:      version,
-	}, nil
-}
-
-// downloadWorkflows downloads all .md files from the workflows directory of a GitHub repository
-func downloadWorkflows(repo, version, targetDir string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Downloading workflows from %s/workflows...\n", repo)
-	}
-
-	// Create a temporary directory for cloning
-	tempDir, err := os.MkdirTemp("", "gh-aw-clone-*")
-	if err != nil {
-		return fmt.Errorf("failed to create temp directory: %w", err)
-	}
-	defer os.RemoveAll(tempDir)
-
-	// Prepare clone arguments - handle SHA commits vs branches/tags differently
-	var cloneArgs []string
-	isSHA := isCommitSHA(version)
-
-	if isSHA {
-		// For commit SHAs, we need full clone to reach the specific commit
-		cloneArgs = []string{"repo", "clone", repo, tempDir}
-	} else {
-		// For branches/tags, use shallow clone for efficiency
-		cloneArgs = []string{"repo", "clone", repo, tempDir, "--", "--depth", "1"}
-		if version != "" && version != "main" {
-			cloneArgs = append(cloneArgs, "--branch", version)
-		}
-	}
-
-	if verbose {
-		fmt.Printf("Cloning repository: gh %s\n", strings.Join(cloneArgs, " "))
-	}
-
-	// Clone the repository
-	_, stdErr, err := gh.Exec(cloneArgs...)
-	if err != nil {
-		return fmt.Errorf("failed to clone repository: %w (stderr: %s)", err, stdErr.String())
-	}
-
-	// If a specific SHA was requested, checkout that commit
-	if isSHA {
-		cmd := exec.Command("git", "checkout", version)
-		cmd.Dir = tempDir
-		if output, err := cmd.CombinedOutput(); err != nil {
-			return fmt.Errorf("failed to checkout commit %s: %w (output: %s)", version, err, string(output))
-		}
-		if verbose {
-			fmt.Printf("Checked out commit: %s\n", version)
-		}
-	}
-
-	// Get the current commit SHA from the cloned repository
-	cmd := exec.Command("git", "rev-parse", "HEAD")
-	cmd.Dir = tempDir
-	commitBytes, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to get commit SHA: %w", err)
-	}
-	commitSHA := strings.TrimSpace(string(commitBytes))
-
-	// Validate that we're at the expected commit if a specific SHA was requested
-	if isSHA && commitSHA != version {
-		return fmt.Errorf("cloned repository is at commit %s, but expected %s", commitSHA, version)
-	}
-
-	if verbose {
-		fmt.Printf("Repository commit SHA: %s\n", commitSHA)
-	}
-
-	// Check if workflows directory exists
-	workflowsDir := filepath.Join(tempDir, "workflows")
-	if _, err := os.Stat(workflowsDir); os.IsNotExist(err) {
-		return fmt.Errorf("workflows directory not found in repository %s", repo)
-	}
-
-	// Copy all .md files from workflows directory to target
-	if err := copyMarkdownFiles(workflowsDir, targetDir, verbose); err != nil {
-		return err
-	}
-
-	// Store the commit SHA in a metadata file
-	metadataFile := filepath.Join(targetDir, ".aw-metadata")
-	metadataContent := fmt.Sprintf("commit_sha=%s\n", commitSHA)
-	if err := os.WriteFile(metadataFile, []byte(metadataContent), 0644); err != nil {
-		return fmt.Errorf("failed to write metadata file: %w", err)
-	}
-
-	if verbose {
-		fmt.Printf("Stored commit SHA in metadata file: %s\n", metadataFile)
-	}
-
-	return nil
-}
-
-// copyMarkdownFiles recursively copies markdown files from source to target directory
-func copyMarkdownFiles(sourceDir, targetDir string, verbose bool) error {
-	return filepath.Walk(sourceDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip if not a markdown file
-		if info.IsDir() || !strings.HasSuffix(info.Name(), ".md") {
-			return nil
-		}
-
-		// Get relative path from source directory
-		relPath, err := filepath.Rel(sourceDir, path)
-		if err != nil {
-			return fmt.Errorf("failed to get relative path: %w", err)
-		}
-
-		// Create target file path
-		targetFile := filepath.Join(targetDir, relPath)
-
-		// Create target directory if needed
-		targetFileDir := filepath.Dir(targetFile)
-		if err := os.MkdirAll(targetFileDir, 0755); err != nil {
-			return fmt.Errorf("failed to create target directory %s: %w", targetFileDir, err)
-		}
-
-		// Copy file
-		if verbose {
-			fmt.Printf("Copying: %s -> %s\n", relPath, targetFile)
-		}
-
-		content, err := os.ReadFile(path)
-		if err != nil {
-			return fmt.Errorf("failed to read source file %s: %w", path, err)
-		}
-
-		if err := os.WriteFile(targetFile, content, 0644); err != nil {
-			return fmt.Errorf("failed to write target file %s: %w", targetFile, err)
-		}
-
-		return nil
-	})
-}
-
-// WorkflowSourceInfo contains information about where a workflow was found
-type WorkflowSourceInfo struct {
-	PackagePath string
-	SourcePath  string
-}
-
-// findWorkflowInPackages searches for a workflow in installed packages
-func findWorkflowInPackages(workflowPath string, verbose bool) ([]byte, *WorkflowSourceInfo, error) {
-	// Check global packages only
-	// Remove .md extension if present for searching
-	workflowName := strings.TrimSuffix(workflowPath, ".md")
-
-	packagesDir, err := getPackagesDir()
-	if err != nil {
-		if verbose {
-			fmt.Printf("Warning: Failed to get packages directory: %v\n", err)
-		}
-		return nil, nil, fmt.Errorf("failed to get packages directory: %w", err)
-	}
-
-	locationName := "global"
-
-	if _, err := os.Stat(packagesDir); os.IsNotExist(err) {
-		if verbose {
-			fmt.Printf("No %s packages directory found at %s\n", locationName, packagesDir)
-		}
-		return nil, nil, fmt.Errorf("no packages directory found")
-	}
-
-	if verbose {
-		fmt.Printf("Searching %s packages in %s for workflow: %s\n", locationName, packagesDir, workflowName)
-	}
-
-	// Check if workflow name contains org/repo prefix
-	if strings.Contains(workflowName, "/") {
-		// Fully qualified name: org/repo/workflow_name
-		content, sourceInfo, err := findQualifiedWorkflowInPackages(workflowName, packagesDir, verbose)
-		if err == nil {
-			return content, sourceInfo, nil
-		}
-		if verbose {
-			fmt.Printf("Qualified workflow not found in %s packages: %v\n", locationName, err)
-		}
-	} else {
-		// Simple name: workflow_name - search all packages
-		content, sourceInfo, err := findUnqualifiedWorkflowInPackages(workflowName, packagesDir, verbose)
-		if err == nil {
-			return content, sourceInfo, nil
-		}
-		if verbose {
-			fmt.Printf("Unqualified workflow not found in %s packages: %v\n", locationName, err)
-		}
-	}
-
-	return nil, nil, fmt.Errorf("workflow not found in components and no packages installed")
-}
-
-// findQualifiedWorkflowInPackages finds a workflow using fully qualified name
-func findQualifiedWorkflowInPackages(qualifiedName, packagesDir string, verbose bool) ([]byte, *WorkflowSourceInfo, error) {
-	parts := strings.Split(qualifiedName, "/")
-	if len(parts) < 3 {
-		return nil, nil, fmt.Errorf("qualified workflow name must be in format 'org/repo/workflow_name'")
-	}
-
-	org := parts[0]
-	repo := parts[1]
-	workflowName := strings.Join(parts[2:], "/") // Support nested workflows
-
-	packagePath := filepath.Join(packagesDir, org, repo)
-	workflowFile := filepath.Join(packagePath, workflowName+".md")
-
-	if verbose {
-		fmt.Printf("Looking for qualified workflow: %s\n", workflowFile)
-	}
-
-	content, err := os.ReadFile(workflowFile)
-	if err != nil {
-		return nil, nil, fmt.Errorf("workflow '%s' not found in package '%s/%s'", workflowName, org, repo)
-	}
-
-	return content, &WorkflowSourceInfo{
-		PackagePath: packagePath,
-		SourcePath:  workflowFile,
-	}, nil
-}
-
-// findUnqualifiedWorkflowInPackages finds a workflow by name across all packages
-func findUnqualifiedWorkflowInPackages(workflowName, packagesDir string, verbose bool) ([]byte, *WorkflowSourceInfo, error) {
-	var matches []WorkflowMatch
-
-	// Search all packages for workflows with this name
-	err := filepath.Walk(packagesDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !info.IsDir() && strings.HasSuffix(info.Name(), ".md") {
-			// Check if this is the workflow we're looking for
-			baseName := strings.TrimSuffix(info.Name(), ".md")
-			if baseName == workflowName {
-				// Get package info from path
-				relPath, err := filepath.Rel(packagesDir, path)
-				if err != nil {
-					return err
-				}
-
-				pathParts := strings.Split(filepath.Dir(relPath), string(filepath.Separator))
-				if len(pathParts) >= 2 {
-					org := pathParts[0]
-					repo := pathParts[1]
-					matches = append(matches, WorkflowMatch{
-						Path:        path,
-						PackageName: fmt.Sprintf("%s/%s", org, repo),
-						Org:         org,
-						Repo:        repo,
-					})
-				}
-			}
-		}
-		return nil
-	})
-
-	if err != nil {
-		return nil, nil, fmt.Errorf("error searching packages: %w", err)
-	}
-
-	if len(matches) == 0 {
-		return nil, nil, fmt.Errorf("workflow '%s' not found in any package", workflowName)
-	}
-
-	if len(matches) == 1 {
-		// Single match, use it
-		match := matches[0]
-		content, err := os.ReadFile(match.Path)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to read workflow file: %w", err)
-		}
-
-		if verbose {
-			fmt.Printf("Found workflow '%s' in package '%s'\n", workflowName, match.PackageName)
-		}
-
-		return content, &WorkflowSourceInfo{
-			PackagePath: filepath.Dir(match.Path),
-			SourcePath:  match.Path,
-		}, nil
-	}
-
-	// Multiple matches - require disambiguation
-	fmt.Printf("Multiple workflows named '%s' found:\n", workflowName)
-	for _, match := range matches {
-		fmt.Printf("  - %s/%s\n", match.PackageName, workflowName)
-	}
-	fmt.Printf("\nPlease specify the full path: "+constants.CLIExtensionPrefix+" add <org/repo/%s>\n", workflowName)
-	fmt.Printf("Or use a different name: "+constants.CLIExtensionPrefix+" add %s -n <custom-name>\n", workflowName)
-
-	return nil, nil, fmt.Errorf("ambiguous workflow name - specify full path or use -n flag for custom name")
-}
-
-// WorkflowMatch represents a workflow match in package search
-type WorkflowMatch struct {
-	Path        string
-	PackageName string
-	Org         string
-	Repo        string
-}
-
-// collectPackageIncludeDependencies collects dependencies for package-based workflows
-func collectPackageIncludeDependencies(content, packagePath string, verbose bool) ([]IncludeDependency, error) {
-	var dependencies []IncludeDependency
-	seen := make(map[string]bool)
-
-	if verbose {
-		fmt.Printf("Collecting package dependencies from: %s\n", packagePath)
-	}
-
-	err := collectPackageIncludesRecursive(content, packagePath, &dependencies, seen, verbose)
-	return dependencies, err
-}
-
-// collectPackageIncludesRecursive recursively processes @include directives in package content
-func collectPackageIncludesRecursive(content, baseDir string, dependencies *[]IncludeDependency, seen map[string]bool, verbose bool) error {
-	includePattern := regexp.MustCompile(`^@include(\?)?\s+(.+)$`)
-
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	for scanner.Scan() {
-		line := scanner.Text()
-		if matches := includePattern.FindStringSubmatch(line); matches != nil {
-			isOptional := matches[1] == "?"
-			includePath := strings.TrimSpace(matches[2])
-
-			// Handle section references (file.md#Section)
-			var filePath string
-			if strings.Contains(includePath, "#") {
-				parts := strings.SplitN(includePath, "#", 2)
-				filePath = parts[0]
-			} else {
-				filePath = includePath
-			}
-
-			// Resolve the full source path relative to base directory
-			fullSourcePath := filepath.Join(baseDir, filePath)
-
-			// Skip if we've already processed this file
-			if seen[fullSourcePath] {
-				continue
-			}
-			seen[fullSourcePath] = true
-
-			// Add dependency
-			dep := IncludeDependency{
-				SourcePath: fullSourcePath,
-				TargetPath: filePath, // Keep relative path for target
-				IsOptional: isOptional,
-			}
-			*dependencies = append(*dependencies, dep)
-
-			if verbose {
-				fmt.Printf("Found include dependency: %s -> %s\n", fullSourcePath, filePath)
-			}
-
-			// Read the included file and process its includes recursively
-			includedContent, err := os.ReadFile(fullSourcePath)
-			if err != nil {
-				if verbose {
-					fmt.Printf("Warning: Could not read include file %s: %v\n", fullSourcePath, err)
-				}
-				continue
-			}
-
-			// Extract markdown content from the included file
-			markdownContent, err := parser.ExtractMarkdownContent(string(includedContent))
-			if err != nil {
-				if verbose {
-					fmt.Printf("Warning: Could not extract markdown from %s: %v\n", fullSourcePath, err)
-				}
-				continue
-			}
-
-			// Recursively process includes in the included file
-			includedDir := filepath.Dir(fullSourcePath)
-			if err := collectPackageIncludesRecursive(markdownContent, includedDir, dependencies, seen, verbose); err != nil {
-				if verbose {
-					fmt.Printf("Warning: Error processing includes in %s: %v\n", fullSourcePath, err)
-				}
-			}
-		}
-	}
-
-	return scanner.Err()
-}
-
-// copyIncludeDependenciesFromPackageWithForce copies include dependencies from package filesystem with force option
-func copyIncludeDependenciesFromPackageWithForce(dependencies []IncludeDependency, githubWorkflowsDir string, verbose bool, force bool, tracker *FileTracker) error {
-	for _, dep := range dependencies {
-		// Create the target path in .github/workflows
-		targetPath := filepath.Join(githubWorkflowsDir, dep.TargetPath)
-
-		// Create target directory if it doesn't exist
-		targetDir := filepath.Dir(targetPath)
-		if err := os.MkdirAll(targetDir, 0755); err != nil {
-			return fmt.Errorf("failed to create directory %s: %w", targetDir, err)
-		}
-
-		// Read source content from package
-		sourceContent, err := os.ReadFile(dep.SourcePath)
-		if err != nil {
-			if dep.IsOptional {
-				// For optional includes, just show an informational message and skip
-				if verbose {
-					fmt.Printf("Optional include file not found: %s (you can create this file to configure the workflow)\n", dep.TargetPath)
-				}
-				continue
-			}
-			fmt.Printf("Warning: Failed to read include file %s: %v\n", dep.SourcePath, err)
-			continue
-		}
-
-		// Check if target file already exists
-		fileExists := false
-		if existingContent, err := os.ReadFile(targetPath); err == nil {
-			fileExists = true
-			// File exists, compare contents
-			if string(existingContent) == string(sourceContent) {
-				// Contents are the same, skip
-				if verbose {
-					fmt.Printf("Include file %s already exists with same content, skipping\n", dep.TargetPath)
-				}
-				continue
-			}
-
-			// Contents are different
-			if !force {
-				fmt.Printf("Include file %s already exists with different content, skipping (use --force to overwrite)\n", dep.TargetPath)
-				continue
-			}
-
-			// Force is enabled, overwrite
-			fmt.Printf("Overwriting existing include file: %s\n", dep.TargetPath)
-		}
-
-		// Track the file based on whether it existed before (if tracker is available)
-		if tracker != nil {
-			if fileExists {
-				tracker.TrackModified(targetPath)
-			} else {
-				tracker.TrackCreated(targetPath)
-			}
-		}
-
-		// Write to target
-		if err := os.WriteFile(targetPath, sourceContent, 0644); err != nil {
-			return fmt.Errorf("failed to write include file %s: %w", targetPath, err)
-		}
-
-		if verbose {
-			fmt.Printf("Copied include file: %s -> %s\n", dep.SourcePath, targetPath)
-		}
-	}
-
-	return nil
-}
-
-// cleanupOrphanedIncludes removes include files that are no longer used by any workflow
-func cleanupOrphanedIncludes(verbose bool) error {
-	// Get all remaining markdown files
-	mdFiles, err := getMarkdownWorkflowFiles()
-	if err != nil {
-		// No markdown files means we can clean up all includes
-		if verbose {
-			fmt.Printf("No markdown files found, cleaning up all includes\n")
-		}
-		return cleanupAllIncludes(verbose)
-	}
-
-	// Collect all include dependencies from remaining workflows
-	usedIncludes := make(map[string]bool)
-
-	for _, mdFile := range mdFiles {
-		content, err := os.ReadFile(mdFile)
-		if err != nil {
-			if verbose {
-				fmt.Printf("Warning: Could not read %s for include analysis: %v\n", mdFile, err)
-			}
-			continue
-		}
-
-		// Find includes used by this workflow
-		includes, err := findIncludesInContent(string(content), filepath.Dir(mdFile), verbose)
-		if err != nil {
-			if verbose {
-				fmt.Printf("Warning: Could not analyze includes in %s: %v\n", mdFile, err)
-			}
-			continue
-		}
-
-		for _, include := range includes {
-			usedIncludes[include] = true
-		}
-	}
-
-	// Find all include files in .github/workflows
-	// Only consider files in subdirectories (like shared/) as potential include files
-	// Root-level .md files are workflow files, not include files
-	workflowsDir := ".github/workflows"
-	var allIncludes []string
-
-	err = filepath.Walk(workflowsDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !info.IsDir() && strings.HasSuffix(info.Name(), ".md") {
-			relPath, err := filepath.Rel(workflowsDir, path)
-			if err != nil {
-				return err
-			}
-
-			// Only consider files in subdirectories as potential include files
-			// Root-level .md files are workflow files, not include files
-			if strings.Contains(relPath, string(filepath.Separator)) {
-				allIncludes = append(allIncludes, relPath)
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to scan include files: %w", err)
-	}
-
-	// Remove unused includes
-	for _, include := range allIncludes {
-		if !usedIncludes[include] {
-			includePath := filepath.Join(workflowsDir, include)
-			if err := os.Remove(includePath); err != nil {
-				if verbose {
-					fmt.Printf("Warning: Failed to remove orphaned include %s: %v\n", include, err)
-				}
-			} else {
-				fmt.Printf("Removed orphaned include: %s\n", include)
-			}
-		}
-	}
-
-	return nil
-}
-
-// previewOrphanedIncludes returns a list of include files that would become orphaned if the specified files were removed
-func previewOrphanedIncludes(filesToRemove []string, verbose bool) ([]string, error) {
-	// Get all current markdown files
-	allMdFiles, err := getMarkdownWorkflowFiles()
-	if err != nil {
-		return nil, err
-	}
-
-	// Create a map of files to remove for quick lookup
-	removeMap := make(map[string]bool)
-	for _, file := range filesToRemove {
-		removeMap[file] = true
-	}
-
-	// Get the files that would remain after removal
-	var remainingFiles []string
-	for _, file := range allMdFiles {
-		if !removeMap[file] {
-			remainingFiles = append(remainingFiles, file)
-		}
-	}
-
-	// If no files remain, all include files would be orphaned
-	if len(remainingFiles) == 0 {
-		return getAllIncludeFiles()
-	}
-
-	// Collect all include dependencies from remaining workflows
-	usedIncludes := make(map[string]bool)
-
-	for _, mdFile := range remainingFiles {
-		content, err := os.ReadFile(mdFile)
-		if err != nil {
-			if verbose {
-				fmt.Printf("Warning: Could not read %s for include analysis: %v\n", mdFile, err)
-			}
-			continue
-		}
-
-		// Find includes used by this workflow
-		includes, err := findIncludesInContent(string(content), filepath.Dir(mdFile), verbose)
-		if err != nil {
-			if verbose {
-				fmt.Printf("Warning: Could not analyze includes in %s: %v\n", mdFile, err)
-			}
-			continue
-		}
-
-		for _, include := range includes {
-			usedIncludes[include] = true
-		}
-	}
-
-	// Find all include files and check which ones would be orphaned
-	allIncludes, err := getAllIncludeFiles()
-	if err != nil {
-		return nil, err
-	}
-
-	var orphanedIncludes []string
-	for _, include := range allIncludes {
-		if !usedIncludes[include] {
-			orphanedIncludes = append(orphanedIncludes, include)
-		}
-	}
-
-	return orphanedIncludes, nil
-}
-
-// getAllIncludeFiles returns all include files in .github/workflows subdirectories
-func getAllIncludeFiles() ([]string, error) {
-	workflowsDir := ".github/workflows"
-	var allIncludes []string
-
-	err := filepath.Walk(workflowsDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !info.IsDir() && strings.HasSuffix(info.Name(), ".md") {
-			relPath, err := filepath.Rel(workflowsDir, path)
-			if err != nil {
-				return err
-			}
-
-			// Only consider files in subdirectories as potential include files
-			// Root-level .md files are workflow files, not include files
-			if strings.Contains(relPath, string(filepath.Separator)) {
-				allIncludes = append(allIncludes, relPath)
-			}
-		}
-
-		return nil
-	})
-
-	return allIncludes, err
-}
-
-// cleanupAllIncludes removes all include files when no workflows remain
-func cleanupAllIncludes(verbose bool) error {
-	workflowsDir := ".github/workflows"
-
-	err := filepath.Walk(workflowsDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !info.IsDir() && strings.HasSuffix(info.Name(), ".md") {
-			relPath, _ := filepath.Rel(workflowsDir, path)
-
-			// Only remove files in subdirectories (like shared/) as these are include files
-			// Root-level .md files are workflow files, not include files
-			if strings.Contains(relPath, string(filepath.Separator)) {
-				if err := os.Remove(path); err != nil {
-					if verbose {
-						fmt.Printf("Warning: Failed to remove include %s: %v\n", relPath, err)
-					}
-				} else {
-					fmt.Printf("Removed include: %s\n", relPath)
-				}
-			}
-		}
-
-		return nil
-	})
-
-	return err
-}
-
-// findIncludesInContent finds all @include references in content
-func findIncludesInContent(content, baseDir string, verbose bool) ([]string, error) {
-	_ = baseDir // unused parameter for now, keeping for potential future use
-	_ = verbose // unused parameter for now, keeping for potential future use
-	var includes []string
-	includePattern := regexp.MustCompile(`^@include(\?)?\s+(.+)$`)
-
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	for scanner.Scan() {
-		line := scanner.Text()
-		if matches := includePattern.FindStringSubmatch(line); matches != nil {
-			includePath := strings.TrimSpace(matches[2])
-
-			// Handle section references (file.md#Section)
-			var filePath string
-			if strings.Contains(includePath, "#") {
-				parts := strings.SplitN(includePath, "#", 2)
-				filePath = parts[0]
-			} else {
-				filePath = includePath
-			}
-
-			includes = append(includes, filePath)
-		}
-	}
-
-	return includes, scanner.Err()
 }
 
 // resolveWorkflowFile resolves a file or workflow name to an actual file path
@@ -2673,604 +86,6 @@ func resolveWorkflowFile(fileOrWorkflowName string, verbose bool) (string, error
 		return path, nil // fallback to original path
 	}
 	return absPath, nil
-}
-
-// RunWorkflowOnGitHub runs an agentic workflow on GitHub Actions
-func RunWorkflowOnGitHub(workflowIdOrName string, enable bool, verbose bool) error {
-	if workflowIdOrName == "" {
-		return fmt.Errorf("workflow name or ID is required")
-	}
-
-	if verbose {
-		fmt.Printf("Running workflow on GitHub Actions: %s\n", workflowIdOrName)
-	}
-
-	// Check if gh CLI is available
-	if !isGHCLIAvailable() {
-		return fmt.Errorf("GitHub CLI (gh) is required but not available")
-	}
-
-	// Try to resolve the workflow file path to find the corresponding .lock.yml file
-	workflowFile, err := resolveWorkflowFile(workflowIdOrName, verbose)
-	if err != nil {
-		return fmt.Errorf("failed to resolve workflow: %w", err)
-	}
-
-	// Check if the workflow is runnable (has workflow_dispatch trigger)
-	runnable, err := IsRunnable(workflowFile)
-	if err != nil {
-		return fmt.Errorf("failed to check if workflow %s is runnable: %w", workflowFile, err)
-	}
-
-	if !runnable {
-		return fmt.Errorf("workflow '%s' cannot be run on GitHub Actions - it must have 'workflow_dispatch' trigger", workflowIdOrName)
-	}
-
-	// Handle --enable flag logic: check workflow state and enable if needed
-	var wasDisabled bool
-	var workflowID int64
-	if enable {
-		// Get current workflow status
-		workflow, err := getWorkflowStatus(workflowIdOrName, verbose)
-		if err != nil {
-			if verbose {
-				fmt.Printf("Warning: Could not check workflow status: %v\n", err)
-			}
-		}
-
-		// If we successfully got workflow status, check if it needs enabling
-		if err == nil {
-			workflowID = workflow.ID
-			if workflow.State == "disabled_manually" {
-				wasDisabled = true
-				if verbose {
-					fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Workflow '%s' is disabled, enabling it temporarily...", workflowIdOrName)))
-				}
-				// Enable the workflow
-				cmd := exec.Command("gh", "workflow", "enable", strconv.FormatInt(workflow.ID, 10))
-				if err := cmd.Run(); err != nil {
-					return fmt.Errorf("failed to enable workflow '%s': %w", workflowIdOrName, err)
-				}
-				fmt.Println(console.FormatSuccessMessage(fmt.Sprintf("Enabled workflow: %s", workflowIdOrName)))
-			}
-		}
-	}
-
-	// Determine the lock file name based on the workflow source
-	var lockFileName string
-
-	// Check that the workflow exists locally before trying to run it
-	workflowsDir := getWorkflowsDir()
-
-	_, _, err = readWorkflowFile(workflowIdOrName+".md", workflowsDir)
-	if err != nil {
-		return fmt.Errorf("failed to find workflow in local .github/workflows or components: %w", err)
-	}
-
-	// For local workflows, use the simple filename
-	filename := strings.TrimSuffix(filepath.Base(workflowIdOrName), ".md")
-	lockFileName = filename + ".lock.yml"
-
-	// Check if the lock file exists in .github/workflows
-	lockFilePath := filepath.Join(".github/workflows", lockFileName)
-	if _, err := os.Stat(lockFilePath); os.IsNotExist(err) {
-		return fmt.Errorf("workflow lock file '%s' not found in .github/workflows - run '"+constants.CLIExtensionPrefix+" compile' first", lockFileName)
-	}
-
-	if verbose {
-		fmt.Printf("Using lock file: %s\n", lockFileName)
-	}
-
-	// Execute gh workflow run command and capture output
-	cmd := exec.Command("gh", "workflow", "run", lockFileName)
-
-	if verbose {
-		fmt.Printf("Executing: gh workflow run %s\n", lockFileName)
-	}
-
-	// Capture both stdout and stderr
-	stdout, err := cmd.Output()
-	if err != nil {
-		// If there's an error, try to get stderr for better error reporting
-		if exitError, ok := err.(*exec.ExitError); ok {
-			fmt.Fprintf(os.Stderr, "%s", exitError.Stderr)
-		}
-
-		// Restore workflow state if it was disabled and we enabled it (even on error)
-		if enable && wasDisabled && workflowID != 0 {
-			restoreWorkflowState(workflowIdOrName, workflowID, verbose)
-		}
-
-		return fmt.Errorf("failed to run workflow on GitHub Actions: %w", err)
-	}
-
-	// Display the output from gh workflow run
-	output := strings.TrimSpace(string(stdout))
-	if output != "" {
-		fmt.Println(output)
-	}
-
-	fmt.Printf("Successfully triggered workflow: %s\n", lockFileName)
-
-	// Try to get the latest run for this workflow to show a direct link
-	// Add a delay to allow GitHub Actions time to register the new workflow run
-	if runInfo, err := getLatestWorkflowRunWithRetry(lockFileName, "", verbose); err == nil && runInfo.URL != "" {
-		fmt.Printf("\n🔗 View workflow run: %s\n", runInfo.URL)
-	} else if verbose && err != nil {
-		fmt.Printf("Note: Could not get workflow run URL: %v\n", err)
-	}
-
-	// Restore workflow state if it was disabled and we enabled it
-	if enable && wasDisabled && workflowID != 0 {
-		restoreWorkflowState(workflowIdOrName, workflowID, verbose)
-	}
-
-	return nil
-}
-
-// RunWorkflowsOnGitHub runs multiple agentic workflows on GitHub Actions, optionally repeating at intervals
-func RunWorkflowsOnGitHub(workflowNames []string, repeatSeconds int, enable bool, verbose bool) error {
-	if len(workflowNames) == 0 {
-		return fmt.Errorf("at least one workflow name or ID is required")
-	}
-
-	// Validate all workflows exist and are runnable before starting
-	for _, workflowName := range workflowNames {
-		if workflowName == "" {
-			return fmt.Errorf("workflow name cannot be empty")
-		}
-
-		// Check if workflow exists and is runnable
-		workflowFile, err := resolveWorkflowFile(workflowName, verbose)
-		if err != nil {
-			return fmt.Errorf("failed to resolve workflow '%s': %w", workflowName, err)
-		}
-
-		runnable, err := IsRunnable(workflowFile)
-		if err != nil {
-			return fmt.Errorf("failed to check if workflow '%s' is runnable: %w", workflowName, err)
-		}
-
-		if !runnable {
-			return fmt.Errorf("workflow '%s' cannot be run on GitHub Actions - it must have 'workflow_dispatch' trigger", workflowName)
-		}
-	}
-
-	// Function to run all workflows once
-	runAllWorkflows := func() error {
-		fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Running %d workflow(s)...", len(workflowNames))))
-
-		for i, workflowName := range workflowNames {
-			if len(workflowNames) > 1 {
-				fmt.Println(console.FormatProgressMessage(fmt.Sprintf("Running workflow %d/%d: %s", i+1, len(workflowNames), workflowName)))
-			}
-
-			if err := RunWorkflowOnGitHub(workflowName, enable, verbose); err != nil {
-				return fmt.Errorf("failed to run workflow '%s': %w", workflowName, err)
-			}
-
-			// Add a small delay between workflows to avoid overwhelming GitHub API
-			if i < len(workflowNames)-1 {
-				time.Sleep(1 * time.Second)
-			}
-		}
-
-		fmt.Println(console.FormatSuccessMessage(fmt.Sprintf("Successfully triggered %d workflow(s)", len(workflowNames))))
-		return nil
-	}
-
-	// Run workflows once
-	if err := runAllWorkflows(); err != nil {
-		return err
-	}
-
-	// If repeat is specified, set up a ticker
-	if repeatSeconds > 0 {
-		fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Repeating every %d seconds. Press Ctrl+C to stop.", repeatSeconds)))
-
-		ticker := time.NewTicker(time.Duration(repeatSeconds) * time.Second)
-		defer ticker.Stop()
-
-		// Set up signal handling for graceful shutdown
-		sigChan := make(chan os.Signal, 1)
-		signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
-
-		for {
-			select {
-			case <-ticker.C:
-				fmt.Println(console.FormatInfoMessage(fmt.Sprintf("Repeating workflow run at %s", time.Now().Format("2006-01-02 15:04:05"))))
-				if err := runAllWorkflows(); err != nil {
-					fmt.Fprintln(os.Stderr, console.FormatErrorMessage(fmt.Sprintf("Error during repeat: %v", err)))
-					// Continue running on error during repeat
-				}
-			case <-sigChan:
-				fmt.Println(console.FormatInfoMessage("Received interrupt signal, stopping repeat..."))
-				return nil
-			}
-		}
-	}
-
-	return nil
-}
-
-// IsRunnable checks if a workflow can be run (has schedule or workflow_dispatch trigger)
-func IsRunnable(markdownPath string) (bool, error) {
-	// Read the file
-	contentBytes, err := os.ReadFile(markdownPath)
-	if err != nil {
-		return false, fmt.Errorf("failed to read file: %w", err)
-	}
-	content := string(contentBytes)
-
-	// Extract frontmatter
-	result, err := parser.ExtractFrontmatterFromContent(content)
-	if err != nil {
-		return false, fmt.Errorf("failed to extract frontmatter: %w", err)
-	}
-
-	// Check if 'on' section is present
-	onSection, exists := result.Frontmatter["on"]
-	if !exists {
-		// If no 'on' section, it defaults to runnable triggers (schedule, workflow_dispatch)
-		return true, nil
-	}
-
-	// Convert to string to analyze
-	onStr := fmt.Sprintf("%v", onSection)
-	onStrLower := strings.ToLower(onStr)
-
-	// Check for schedule or workflow_dispatch triggers
-	hasSchedule := strings.Contains(onStrLower, "schedule") || strings.Contains(onStrLower, "cron")
-	hasWorkflowDispatch := strings.Contains(onStrLower, "workflow_dispatch")
-
-	return hasSchedule || hasWorkflowDispatch, nil
-}
-
-// findMatchingLockFile searches for existing lock files that might match the given workflow name
-func findMatchingLockFile(workflowName string, verbose bool) string {
-	workflowsDir := getWorkflowsDir()
-
-	// Look for any .lock.yml files that might correspond to this workflow
-	lockFiles, err := filepath.Glob(filepath.Join(workflowsDir, "*.lock.yml"))
-	if err != nil {
-		if verbose {
-			fmt.Printf("Warning: Failed to search for lock files: %v\n", err)
-		}
-		return ""
-	}
-
-	if verbose {
-		fmt.Printf("Searching for lock files matching workflow '%s'\n", workflowName)
-	}
-
-	// Look for exact matches first, then partial matches
-	for _, lockFile := range lockFiles {
-		baseName := filepath.Base(lockFile)
-		lockName := strings.TrimSuffix(baseName, ".lock.yml")
-
-		// Check if the lock file ends with the workflow name (for qualified names)
-		if strings.HasSuffix(lockName, "_"+workflowName) {
-			if verbose {
-				fmt.Printf("Found matching lock file (suffix match): %s\n", baseName)
-			}
-			return baseName
-		}
-	}
-
-	// If no suffix match, look for any lock file containing the workflow name
-	for _, lockFile := range lockFiles {
-		baseName := filepath.Base(lockFile)
-		lockName := strings.TrimSuffix(baseName, ".lock.yml")
-
-		if strings.Contains(lockName, workflowName) {
-			if verbose {
-				fmt.Printf("Found matching lock file (contains match): %s\n", baseName)
-			}
-			return baseName
-		}
-	}
-
-	if verbose {
-		fmt.Printf("No matching lock file found for workflow '%s'\n", workflowName)
-	}
-	return ""
-}
-
-// WorkflowRunInfo contains information about a workflow run
-type WorkflowRunInfo struct {
-	URL        string
-	DatabaseID int64
-	Status     string
-	Conclusion string
-	CreatedAt  time.Time
-}
-
-// getLatestWorkflowRunWithRetry gets information about the most recent run of the specified workflow
-// with retry logic to handle timing issues when a workflow has just been triggered
-func getLatestWorkflowRunWithRetry(lockFileName string, repo string, verbose bool) (*WorkflowRunInfo, error) {
-	const maxRetries = 6
-	const initialDelay = 2 * time.Second
-	const maxDelay = 10 * time.Second
-
-	if verbose {
-		if repo != "" {
-			fmt.Printf("Getting latest run for workflow: %s in repo: %s (with retry logic)\n", lockFileName, repo)
-		} else {
-			fmt.Printf("Getting latest run for workflow: %s (with retry logic)\n", lockFileName)
-		}
-	}
-
-	// Capture the current time before we start polling
-	// This helps us identify runs that were created after the workflow was triggered
-	startTime := time.Now().UTC()
-
-	var lastErr error
-	for attempt := 0; attempt < maxRetries; attempt++ {
-		if attempt > 0 {
-			// Calculate delay with exponential backoff, capped at maxDelay
-			delay := time.Duration(attempt) * initialDelay
-			if delay > maxDelay {
-				delay = maxDelay
-			}
-
-			if verbose {
-				fmt.Printf("Waiting %v before retry attempt %d/%d...\n", delay, attempt+1, maxRetries)
-			} else if attempt == 1 {
-				// Show spinner only starting from second attempt to avoid flickering
-				spinner := console.NewSpinner("Waiting for workflow run to appear...")
-				spinner.Start()
-				time.Sleep(delay)
-				spinner.Stop()
-				continue
-			}
-			time.Sleep(delay)
-		}
-
-		// Build command with optional repo parameter
-		var cmd *exec.Cmd
-		if repo != "" {
-			cmd = exec.Command("gh", "run", "list", "--repo", repo, "--workflow", lockFileName, "--limit", "1", "--json", "url,databaseId,status,conclusion,createdAt")
-		} else {
-			cmd = exec.Command("gh", "run", "list", "--workflow", lockFileName, "--limit", "1", "--json", "url,databaseId,status,conclusion,createdAt")
-		}
-
-		output, err := cmd.Output()
-		if err != nil {
-			lastErr = fmt.Errorf("failed to get workflow runs: %w", err)
-			if verbose {
-				fmt.Printf("Attempt %d/%d failed: %v\n", attempt+1, maxRetries, err)
-			}
-			continue
-		}
-
-		if len(output) == 0 || string(output) == "[]" {
-			lastErr = fmt.Errorf("no runs found for workflow")
-			if verbose {
-				fmt.Printf("Attempt %d/%d: no runs found yet\n", attempt+1, maxRetries)
-			}
-			continue
-		}
-
-		// Parse the JSON output
-		var runs []struct {
-			URL        string `json:"url"`
-			DatabaseID int64  `json:"databaseId"`
-			Status     string `json:"status"`
-			Conclusion string `json:"conclusion"`
-			CreatedAt  string `json:"createdAt"`
-		}
-
-		if err := json.Unmarshal(output, &runs); err != nil {
-			lastErr = fmt.Errorf("failed to parse workflow run data: %w", err)
-			if verbose {
-				fmt.Printf("Attempt %d/%d failed to parse JSON: %v\n", attempt+1, maxRetries, err)
-			}
-			continue
-		}
-
-		if len(runs) == 0 {
-			lastErr = fmt.Errorf("no runs found")
-			if verbose {
-				fmt.Printf("Attempt %d/%d: no runs in parsed JSON\n", attempt+1, maxRetries)
-			}
-			continue
-		}
-
-		run := runs[0]
-
-		// Parse the creation timestamp
-		var createdAt time.Time
-		if run.CreatedAt != "" {
-			if parsedTime, err := time.Parse(time.RFC3339, run.CreatedAt); err == nil {
-				createdAt = parsedTime
-			} else if verbose {
-				fmt.Printf("Warning: Could not parse creation time '%s': %v\n", run.CreatedAt, err)
-			}
-		}
-
-		runInfo := &WorkflowRunInfo{
-			URL:        run.URL,
-			DatabaseID: run.DatabaseID,
-			Status:     run.Status,
-			Conclusion: run.Conclusion,
-			CreatedAt:  createdAt,
-		}
-
-		// If we found a run and it was created after we started (within 30 seconds tolerance),
-		// it's likely the run we just triggered
-		if !createdAt.IsZero() && createdAt.After(startTime.Add(-30*time.Second)) {
-			if verbose {
-				fmt.Printf("Found recent run (ID: %d) created at %v (started polling at %v)\n",
-					run.DatabaseID, createdAt.Format(time.RFC3339), startTime.Format(time.RFC3339))
-			}
-			return runInfo, nil
-		}
-
-		if verbose {
-			if createdAt.IsZero() {
-				fmt.Printf("Attempt %d/%d: Found run (ID: %d) but no creation timestamp available\n", attempt+1, maxRetries, run.DatabaseID)
-			} else {
-				fmt.Printf("Attempt %d/%d: Found run (ID: %d) but it was created at %v (too old)\n",
-					attempt+1, maxRetries, run.DatabaseID, createdAt.Format(time.RFC3339))
-			}
-		}
-
-		// For the first few attempts, if we have a run but it's too old, keep trying
-		if attempt < 3 {
-			lastErr = fmt.Errorf("workflow run appears to be from a previous execution")
-			continue
-		}
-
-		// For later attempts, return what we found even if timing is uncertain
-		if verbose {
-			fmt.Printf("Returning workflow run (ID: %d) after %d attempts (timing uncertain)\n", run.DatabaseID, attempt+1)
-		}
-		return runInfo, nil
-	}
-
-	// If we exhausted all retries, return the last error
-	if lastErr != nil {
-		return nil, fmt.Errorf("failed to get workflow run after %d attempts: %w", maxRetries, lastErr)
-	}
-
-	return nil, fmt.Errorf("no workflow run found after %d attempts", maxRetries)
-}
-
-// checkCleanWorkingDirectory checks if there are uncommitted changes
-func checkCleanWorkingDirectory(verbose bool) error {
-	if verbose {
-		fmt.Printf("Checking for uncommitted changes...\n")
-	}
-
-	cmd := exec.Command("git", "status", "--porcelain")
-	output, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to check git status: %w", err)
-	}
-
-	if len(strings.TrimSpace(string(output))) > 0 {
-		return fmt.Errorf("working directory has uncommitted changes, please commit or stash them first")
-	}
-
-	if verbose {
-		fmt.Printf("Working directory is clean\n")
-	}
-	return nil
-}
-
-// getCurrentBranch gets the current git branch name
-func getCurrentBranch() (string, error) {
-	cmd := exec.Command("git", "branch", "--show-current")
-	output, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("failed to get current branch: %w", err)
-	}
-
-	branch := strings.TrimSpace(string(output))
-	if branch == "" {
-		return "", fmt.Errorf("could not determine current branch")
-	}
-
-	return branch, nil
-}
-
-// createAndSwitchBranch creates a new branch and switches to it
-func createAndSwitchBranch(branchName string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Creating and switching to branch: %s\n", branchName)
-	}
-
-	cmd := exec.Command("git", "checkout", "-b", branchName)
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to create and switch to branch %s: %w", branchName, err)
-	}
-
-	return nil
-}
-
-// switchBranch switches to the specified branch
-func switchBranch(branchName string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Switching to branch: %s\n", branchName)
-	}
-
-	cmd := exec.Command("git", "checkout", branchName)
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to switch to branch %s: %w", branchName, err)
-	}
-
-	return nil
-}
-
-// commitChanges commits all staged changes with the given message
-func commitChanges(message string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Committing changes with message: %s\n", message)
-	}
-
-	cmd := exec.Command("git", "commit", "-m", message)
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to commit changes: %w", err)
-	}
-
-	return nil
-}
-
-// pushBranch pushes the specified branch to origin
-func pushBranch(branchName string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Pushing branch: %s\n", branchName)
-	}
-
-	cmd := exec.Command("git", "push", "-u", "origin", branchName)
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to push branch %s: %w", branchName, err)
-	}
-
-	return nil
-}
-
-// createPR creates a pull request using GitHub CLI
-func createPR(branchName, title, body string, verbose bool) error {
-	if verbose {
-		fmt.Printf("Creating PR: %s\n", title)
-	}
-
-	// Get the current repository info to ensure PR is created in the correct repo
-	cmd := exec.Command("gh", "repo", "view", "--json", "owner,name")
-	repoOutput, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to get current repository info: %w", err)
-	}
-
-	var repoInfo struct {
-		Owner struct {
-			Login string `json:"login"`
-		} `json:"owner"`
-		Name string `json:"name"`
-	}
-
-	if err := json.Unmarshal(repoOutput, &repoInfo); err != nil {
-		return fmt.Errorf("failed to parse repository info: %w", err)
-	}
-
-	repoSpec := fmt.Sprintf("%s/%s", repoInfo.Owner.Login, repoInfo.Name)
-
-	// Explicitly specify the repository to ensure PR is created in the current repo (not upstream)
-	cmd = exec.Command("gh", "pr", "create", "--repo", repoSpec, "--title", title, "--body", body, "--head", branchName)
-	output, err := cmd.Output()
-	if err != nil {
-		// Try to get stderr for better error reporting
-		if exitError, ok := err.(*exec.ExitError); ok {
-			return fmt.Errorf("failed to create PR: %w\nOutput: %s\nError: %s", err, string(output), string(exitError.Stderr))
-		}
-		return fmt.Errorf("failed to create PR: %w", err)
-	}
-
-	prURL := strings.TrimSpace(string(output))
-	fmt.Printf("📢 Pull Request created: %s\n", prURL)
-
-	return nil
 }
 
 // NewWorkflow creates a new workflow markdown file with template content
